% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_df.R
\name{as.casDataFrame}
\alias{as.casDataFrame}
\title{Convert an \R Data Frame to a CAS Data Frame}
\usage{
as.casDataFrame(
  df,
  name = "",
  label = "",
  title = "",
  attrs = list(),
  col.labels = "",
  col.formats = "",
  col.attrs = list(),
  col.sizes = list(),
  col.types = "",
  col.widths = 0
)
}
\arguments{
\item{df}{The data.frame to encapsulate in the casDataFrame.}

\item{name}{An optional \code{character} string that specifies CAS metadata
for a table name.}

\item{label}{An optional \code{character} string that specifies CAS metadata
for a table label.}

\item{title}{An optional \code{character} string that specifies CAS metadata
for a table title.}

\item{attrs}{An optional \code{list} of key-value pairs the specify
user-defined CAS metadata.}

\item{col.labels}{An optional \code{character} string that specifies CAS
metadata for column labels.}

\item{col.formats}{An optional \code{character} string that specifies CAS
metadata for column formats.}

\item{col.attrs}{An optional \code{list} that specifies CAS metadata for
column attributes.}

\item{col.sizes}{An optional \code{list} that specifies CAS metadata for
the number of bytes in the widest row.}

\item{col.types}{An optional \code{character} that specifies CAS metadata
for the column data types.}

\item{col.widths}{An optional \code{numeric} that specifies CAS metadata
for column widths.}
}
\value{
casDataFrame
}
\description{
This function is rarely used for programming. It is
used by the package to associate CAS metadata with
tabular data that is returned by CAS actions.
}
\examples{
\dontrun{
cdf2 <- as.CASDataFrame(df3[1:4])
cdf <- as.CASDataFrame(iris)
}

}
