% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{as.CASTable,CAS-method}
\alias{as.CASTable,CAS-method}
\title{Upload an Object to a CAS Table}
\usage{
\S4method{as.CASTable}{CAS}(x, frame, casOut = "")
}
\arguments{
\item{frame}{A \code{data.frame} object with the data to
upload to CAS.}

\item{casOut}{An optional \code{character} or list. If
you specify a string, then the string is used as the
in-memory table name. A list can be used to specify
properties for the in-memory table as follows:
\describe{
  \item{\code{name}}{An optional \code{character} that
    specifies the name for the in-memory table. By
    default, the name of the data frame is used.}
  \item{\code{caslib}}{An optional \code{character} that
    specifies the caslib. Specify this parameter to
    override the active caslib.}
  \item{\code{label}}{An optional \code{character} that
    specifies a descriptive label for the data.}
  \item{\code{replace}}{An optional \code{logical}. When
    set to TRUE, you can replace an existing in-memory
    table with the same name in the same caslib. The
    default value is FALSE.}
  \item{\code{promote}}{An optional \code{logical}. When
    set to TRUE, the in-memory table has global scope and
    can be available to other CAS sessions (subject to
    access controls). The default value is FALSE and
    the in-memory table has session scope so that it is
    accessible with the session that uploaded the table
    only. Session-scope tables are ideal for data analysis.
    Global-scope tables are better suited for reporting.}
  \item{\code{replication}}{An optional \code{numeric} that
    specifies the number of redundant copies of in-memory
    blocks. This parameter applies to distributed servers
    only. The default value is 1.}
 }}

\item{conn}{A \code{\link{CAS}} object that represents
a connection and session in CAS.}
}
\value{
\code{\link{CASTable}}
}
\description{
Uploads an \R data frame to CAS and returns a
\code{\link{CASTable}} object. The CASTable object
is a reference in \R (the client) to the in-memory
table that is in CAS (the server).
}
\examples{
\dontrun{
s <- CAS("cloud.example.com", 5570)
irisct <- as.CASTable(s, iris)

# Specify a name for the in-memory table.
mtcarsct <- as.CASTable(s, mtcars, casOut = "mtcarsct")

# Avoid replacing an existing in-memory table.
mtcarsct <- as.CASTable(s, mtcars, casOut = list(name = "mtcarsct", replace = FALSE))
}

}
