% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{as.data.frame,CASTable-method}
\alias{as.data.frame,CASTable-method}
\title{Convert a CAS Table to a Data Frame (Download)}
\usage{
\S4method{as.data.frame}{CASTable}(
  x,
  max.rows = getOption("cas.max.download.rows"),
  sample = FALSE,
  sample.seed = 0,
  ...
)
}
\arguments{
\item{x}{The CAS table data to download.}

\item{obs}{Optional integer indicating the maximum number of rows
of data to download.}
}
\value{
Returns a data.frame object that contains
        a copy of the in-memory data.
}
\description{
Downloads the in-memory table that is referenced by
the CASTable object and stores it as a data.frame
in R. This function is used primarily by the package
to store the results of a CAS action.
}
\examples{
\dontrun{
cdf <- as.data.frame(CASTable)
}

}
