% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{as.view,CASTable-method}
\alias{as.view,CASTable-method}
\title{Create a server-side view using the CASTable attributes}
\usage{
\S4method{as.view}{CASTable}(x)
}
\arguments{
\item{x}{\code{CASTable} object.}
}
\value{
\code{CASTable} object referencing the view.
}
\description{
Create a server-side view using the CASTable attributes
}
