% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaSimulation.R
\name{behaSimulation}
\alias{behaSimulation}
\title{Find behavioral simulation statistics}
\usage{
behaSimulation(
  objValue,
  simData,
  parameterValue,
  behThreshold,
  varNumber,
  statIndex,
  observedData,
  minOrmax,
  samplingApproach
)
}
\arguments{
\item{objValue}{numeric vector of the objective function values}

\item{simData}{A list of list object containing simulated data, the first key
is the variable number and the second key is the simulation number, for example
\code{simData[[variable_number]][[simulation_number]]}}

\item{parameterValue}{a dataframe contains all parametersets for running SWAT
or SWAT+}

\item{behThreshold}{the behavioral threshold}

\item{varNumber}{the index of the variable number}

\item{statIndex}{model performance index, for example:
"NSE",
"KGE",
"R2",
"aBIAS",
"RMSE"}

\item{observedData}{list of observed data, each observed data is a data frame}

\item{minOrmax}{character, "Maximize" or "Minimize" the objective function
should be maximized/minimized. So it depends on the objective (statIndex)
function used in here, for example:
 NSE: "Maximize"
 KGE: "Maximize"
 R2: "Maximize"
 aBIAS: "Minimize"
 RMSE: "Minimize"}

\item{samplingApproach}{the parameter sampling approach, possible values are:
'Sensi_Cali_(uniform_Latin_Hypercube_Sampling)' \cr
'Cali_(from_optimization_package)'\cr
'Cali_(from_nloptr_package)')\cr
'Cali_(Dynamically_Dimensioned_Search)'\cr
'Cali_(Generalized_Likelihood_Uncertainty_Estimation)'\cr
'Read_User_Parameter_File'\cr
'Sensi_(from_userDefined_package)'\cr
'Cali_(from_userDefined_package)'}
}
\value{
a list of dataframes showing the 95PPU, median,
p-factor, and r-factor.
}
\description{
Calculate the median, 95PPU, p-factor, r-factor of simulated variables and
parameters
}
\examples{

\donttest{
# Please see RSWAT Vignettes
}


}
