% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bygroups.R
\name{bygroups.as.columns,data.frame-method}
\alias{bygroups.as.columns,data.frame-method}
\title{Return data.frame with By group values prepended as columns}
\usage{
\S4method{bygroups.as.columns}{data.frame}(x)
}
\arguments{
\item{x}{data.frame object.}
}
\value{
data.frame
}
\description{
Return data.frame with By group values prepended as columns
}
