% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bygroups.R
\name{byvals,data.frame-method}
\alias{byvals,data.frame-method}
\title{Retrieve a vector of the By values}
\usage{
\S4method{byvals}{data.frame}(x)
}
\arguments{
\item{x}{data.frame object.}
}
\value{
list()
}
\description{
Retrieve a vector of the By values
}
