% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bygroups.R
\name{byvars,data.frame-method}
\alias{byvars,data.frame-method}
\title{Retrieve a vector of the By values}
\usage{
\S4method{byvars}{data.frame}(x)
}
\arguments{
\item{x}{data.frame object.}
}
\value{
character vector
}
\description{
Retrieve a vector of the By values
}
