% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calObjFunction.R
\name{calObjFunction}
\alias{calObjFunction}
\title{Read output data from R-SWAT and calculate objective functions for
calibration and validation period and for each variable}
\usage{
calObjFunction(
  parameterValue,
  ncores,
  nOutputVar,
  userReadSwatOutput,
  observedData,
  workingDirectory,
  index
)
}
\arguments{
\item{parameterValue}{a dataframe contains all parametersets for running SWAT
or SWAT+}

\item{ncores}{number of threads for parallel execution}

\item{nOutputVar}{number of output variables}

\item{userReadSwatOutput}{TODO}

\item{observedData}{list of observed data, each observed data is a data frame}

\item{workingDirectory}{path to the working directory where all files created
by SWAT or SWAT+ will be saved}

\item{index}{performance criteria (NSE, KGE, RMSE, R2, or aBIAS)}
}
\value{
dataframe of objective function values
}
\description{
Read output data from R-SWAT and calculate objective functions for
calibration and validation period and for each variable
}
\examples{

# Please see RSWAT vignettes "SUFI2_without_GUI"

}
