% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.close,CAS-method}
\alias{cas.close,CAS-method}
\title{Close a CAS connection while leaving the session alive}
\usage{
\S4method{cas.close}{CAS}(x, close.session = FALSE)
}
\arguments{
\item{x}{The \code{\link{CAS}} connection object}

\item{close.session}{Should the session be terminated in addition to 
closing the connection?}
}
\description{
Close a CAS connection while leaving the session alive
}
\seealso{
\code{\link{cas.terminate}}
}
