% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.close}
\alias{cas.close}
\title{Close a CAS connection while leaving the session alive}
\usage{
cas.close(conn, close_session = FALSE)
}
\arguments{
\item{conn}{The \code{\link{CAS}} connection object}

\item{close_session}{Should the session be terminated in addition to 
closing the connection?}
}
\description{
Close a CAS connection while leaving the session alive
}
\seealso{
\code{\link{cas.terminate}}
}
