% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.copy,CAS-method}
\alias{cas.copy,CAS-method}
\title{Return a copy of the CAS connection object}
\usage{
\S4method{cas.copy}{CAS}(x)
}
\arguments{
\item{x}{\code{CAS} connection object.}
}
\value{
\code{CAS} connection object
}
\description{
The new connection will use all of the same settings as the original,
but it will be attached to a new session.
}
