% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.copy,CASTable-method}
\alias{cas.copy,CASTable-method}
\title{Return a copy of the CASTable object}
\usage{
\S4method{cas.copy}{CASTable}(x)
}
\arguments{
\item{x}{\code{CASTable} object.}
}
\value{
\code{CASTable} object
}
\description{
A new \code{CASTable} object will be created that references the
same table in the server as the original \code{CASTable} object.
}
\details{
Note that this does *not* create a new table on the server.
}
