% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.count,CASTable-method}
\alias{cas.count,CASTable-method}
\title{Count of Nonmissing Values}
\usage{
\S4method{cas.count}{CASTable}(x)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}
}
\value{
Named numeric vector
}
\description{
Returns the number of nonmissing values in the input
table by column.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.count(ct[1:4])
cas.count(ct$n2)
}

}
\seealso{
\code{cas.nmiss} to count missing values.
}
