% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.count}
\alias{cas.count}
\title{Count of Nonmissing Values}
\usage{
cas.count(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
casDataFrame

The result includes one row for each numeric variable
and a column that is named N for the nonmissing count.
}
\description{
Returns the number of nonmissing values in the input
table by column.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.count(ct[1:4])
cas.count(ct$n2)
}
}
\seealso{
\code{cas.nmiss} to count missing values.
}
% Copyright SAS Institute
