% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.css,CASTable-method}
\alias{cas.css,CASTable-method}
\title{Corrected Sum of Squares}
\usage{
\S4method{cas.css}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the corrected sum of squares of the
values for each column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.css(irisct)
x["Sepal.Length"]
x[1:2]
}

}
