% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.css}
\alias{cas.css}
\title{Corrected Sum of Squares}
\usage{
cas.css(x)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}
}
\value{
A named numeric vector. You can access the corrected
  sum of squares by column name or index.
}
\description{
Returns the corrected sum of squares of the
values for each column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.css(irisct)
x["Sepal.Length"]
x[1:2]
}
}
% Copyright SAS Institute
