% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.cv}
\alias{cas.cv}
\title{Coefficient of Variation}
\usage{
cas.cv(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
vector

The result is a named numeric vector. You can access
the coefficient of variation by column name or index.
}
\description{
Returns the coefficient of variation of the 
values for each column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.cv(irisct)
x['Sepal.Length']
x[1:2]
}
}
% Copyright SAS Institute
