% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.debug,CAS-method}
\alias{cas.debug,CAS-method}
\title{Return resulting debug messages of last CAS action}
\usage{
\S4method{cas.debug}{CAS}(.x)
}
\arguments{
\item{.x}{\code{\link{CAS}} connection object}
}
\value{
character
}
\description{
Return resulting debug messages of last CAS action
}
