% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.debug,CASTable-method}
\alias{cas.debug,CASTable-method}
\title{Return resulting debug messages of last CAS action}
\usage{
\S4method{cas.debug}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
character
}
\description{
Return resulting debug messages of last CAS action
}
