% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.debug,list-method}
\alias{cas.debug,list-method}
\title{Return resulting debug messages from result set}
\usage{
\S4method{cas.debug}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
character
}
\description{
Return resulting debug messages from result set
}
