% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.events,CASTable-method}
\alias{cas.events,CASTable-method}
\title{Return events from the last CAS action}
\usage{
\S4method{cas.events}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
list
}
\description{
Events are things such as caslibs or tables being added,
updated, or deleted.
}
