% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.events,list-method}
\alias{cas.events,list-method}
\title{Return events from result set}
\usage{
\S4method{cas.events}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
list
}
\description{
Events are things such as caslibs or tables being added,
updated, or deleted.
}
