% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.fork,CAS-method}
\alias{cas.fork,CAS-method}
\title{Return a vector of \code{num} CAS connection objects}
\usage{
\S4method{cas.fork}{CAS}(x, num = 2)
}
\arguments{
\item{num}{Integer indicating the number of returned objects.}

\item{conn}{CAS connection object.}
}
\value{
Vector of CAS connection objects
}
\description{
The returned vector will always contain the current CAS connection
object as the first item.
}
