% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.help,CAS-method}
\alias{cas.help,CAS-method}
\title{Display help for a given action}
\usage{
\S4method{cas.help}{CAS}(.x, .action)
}
\arguments{
\item{.x}{\code{\link{CAS}} connection object}

\item{.action}{The action name to display help for.}
}
\description{
Display help for a given action
}
