% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.invoke,CAS-method}
\alias{cas.invoke,CAS-method}
\title{Invoke an action on the CAS server and return immediately}
\usage{
\S4method{cas.invoke}{CAS}(.x, .action, ...)
}
\arguments{
\item{.x}{\code{CAS} connection object.}

\item{.action}{String containing action name.}

\item{\ldots}{Action parameters.}
}
\value{
CAS connection object
}
\description{
The \code{.action} parameter is a string containing the action
name. All other arguments are passed to the CAS action.
}
\details{
This method is only called in special circumstances where you want
to run an action in the background and retrieve the responses
manually. Typically, the generated action wrappers 
(e.g., \code{cas.table.columninfo()}, \code{cas.simple.summary()},
and so on) are used to make action calls.
}
