% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.max,CASTable-method}
\alias{cas.max,CASTable-method}
\title{Maximum Values}
\usage{
\S4method{cas.max}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the maximum value for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.max(ct[1:4])
cas.max(ct$n2)
}

}
\seealso{
\code{max,CASTable-method}
}
