% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.max}
\alias{cas.max}
\title{Maximum Values}
\usage{
cas.max(x)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}
}
\value{
casDataFrame
}
\description{
Returns the maximum value for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.max(ct[1:4])
cas.max(ct$n2)
}
}
\seealso{
\code{max,CASTable-method}

The result includes one row for each numeric variable
and a column that is named Max for the maximum value.
}
% Copyright SAS Institute
