% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.mean,CASTable-method}
\alias{cas.mean,CASTable-method}
\title{Average Values}
\usage{
\S4method{cas.mean}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the mean value for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.mean(ct[1:4])
cas.mean(ct$n2)
}

}
\seealso{
\code{mean,CASTable-method}
}
