% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.mean}
\alias{cas.mean}
\title{Average Values}
\usage{
cas.mean(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
casDataFrame

The result includes one row for each numeric variable
and a column that is named Mean for the mean value.
}
\description{
Returns the mean value for each column in 
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.mean(ct[1:4])
cas.mean(ct$n2)
}
}
\seealso{
\code{mean,CASTable-method}
}
% Copyright SAS Institute
