% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.median}
\alias{cas.median}
\title{Median Values}
\usage{
cas.median(x)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}
}
\value{
data.frame
}
\description{
Returns the median value for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.median(ct[1:4])
cas.median(ct$n2)
}
}
\seealso{
\code{median,CASTable-method}

The result includes one row for each numeric variable
and a column that is named Median for the median value.
}
% Copyright SAS Institute
