% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.messages,CAS-method}
\alias{cas.messages,CAS-method}
\title{Return messages from the last CAS action}
\usage{
\S4method{cas.messages}{CAS}(.x)
}
\arguments{
\item{.x}{\code{\link{CAS}} connection object}
}
\value{
character
}
\description{
Return messages from the last CAS action
}
