% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.messages,CASTable-method}
\alias{cas.messages,CASTable-method}
\title{Return messages from the last CAS action}
\usage{
\S4method{cas.messages}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
character
}
\description{
Return messages from the last CAS action
}
