% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.messages,list-method}
\alias{cas.messages,list-method}
\title{Return messages from result set}
\usage{
\S4method{cas.messages}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
character
}
\description{
Return messages from result set
}
