% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.min,CASTable-method}
\alias{cas.min,CASTable-method}
\title{Minimum Values}
\usage{
\S4method{cas.min}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the minimum value for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.min(ct[1:4])
cas.min(ct$n2)
}

}
\seealso{
\code{min,CASTable-method}
}
