% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.min}
\alias{cas.min}
\title{Minimum Values}
\usage{
cas.min(x)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}
}
\value{
A casDataFrame which includes one row for each numeric variable
  and a column that is named Minimum for the minimum value.
}
\description{
Returns the minimum value for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.min(ct[1:4])
cas.min(ct$n2)
}
}
\seealso{
\code{min,CASTable-method}
}
% Copyright SAS Institute
