% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.mode,CASTable-method}
\alias{cas.mode,CASTable-method}
\title{Mode Value}
\usage{
\S4method{cas.mode}{CASTable}(x, max.tie = 25, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{max.tie}{Maximum number of duplicate values to return if values have the same rank.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
data.frame
}
\description{
Returns the value that occurs most often for each column in
the input table and the count for the value.
}
\examples{
\dontrun{
cas.mode(ct[1:4])
cas.mode(ct$n2)
}

}
