% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.mode}
\alias{cas.mode}
\title{Mode Value}
\usage{
cas.mode(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
data.frame

The result includes one row for each variable.
One column is named Mode for the most common value. 
Another column is named Count to show the number of rows
with the most common value.
}
\description{
Returns the value that occurs most often for each column in 
the input table and the count for the value.
}
\details{
This function operates on numeric and character columns.
}
\examples{
\dontrun{
cas.mode(ct[1:4])
cas.mode(ct$n2)
}
}
% Copyright SAS Institute
