% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.nmiss}
\alias{cas.nmiss}
\title{Number of Missing Values}
\usage{
cas.nmiss(x)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}
}
\value{
A named numeric vector. You can access the count of
  missing values by column name or index.
}
\description{
Returns the number of missing values in the input
table by column.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.nmiss(irisct)
x["Sepal.Length"]
x[1:2]
}
}
\seealso{
\code{cas.count} to count nonmissing values.
}
% Copyright SAS Institute
