% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.performance,CAS-method}
\alias{cas.performance,CAS-method}
\title{Return performance metrics for last CAS action}
\usage{
\S4method{cas.performance}{CAS}(.x)
}
\arguments{
\item{.x}{\code{\link{CAS}} connection object}
}
\value{
list
}
\description{
Return performance metrics for last CAS action
}
