% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.performance,CASTable-method}
\alias{cas.performance,CASTable-method}
\title{Return performance metrics for last CAS action}
\usage{
\S4method{cas.performance}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
list
}
\description{
Return performance metrics for last CAS action
}
