% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.performance,list-method}
\alias{cas.performance,list-method}
\title{Return performance metrics from result set}
\usage{
\S4method{cas.performance}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
list
}
\description{
Return performance metrics from result set
}
