% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.probt,CASTable-method}
\alias{cas.probt,CASTable-method}
\title{P-Value of the T-Statistics}
\usage{
\S4method{cas.probt}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the p-values for the
values for each column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.probt(irisct)
x["Sepal.Length"]
x[1:2]
}

}
