% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.probt}
\alias{cas.probt}
\title{P-Value of the T-Statistics}
\usage{
cas.probt(x)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}
}
\value{
A named numeric vector. You can access the p-value by
  column name or index.
}
\description{
Returns the p-values for the
values for each column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.probt(irisct)
x["Sepal.Length"]
x[1:2]
}
}
% Copyright SAS Institute
