% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.quantile,CASTable-method}
\alias{cas.quantile,CASTable-method}
\title{Quantile and Percentile Values}
\usage{
\S4method{cas.quantile}{CASTable}(x, q = seq(0, 100, 25), na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{q}{A list of numeric values.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
data.frame with row labels as variable names and column
        labels as percentage labels (e.g., '25%', '50%', etc.).
}
\description{
Returns the requested percentiles for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.quantile(ct[1:4], q = 50)
cas.quantile(ct$n2, q = c(10, 25, 50, 75, 90))
}

}
