% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.quantile}
\alias{cas.quantile}
\title{Quantile and Percentile Values}
\usage{
cas.quantile(CASTable, q)
}
\arguments{
\item{q}{A list of numeric values.}

\item{x}{CASTable.}
}
\value{
data.frame

The result includes one row for the variable, the requested
percentile, and the value.
}
\description{
Returns the requested percentiles for each column in 
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.quantile(ct[1:4], q=50)
cas.quantile(ct$n2, q=c(10, 25, 50, 75, 90))
}
}
% Copyright SAS Institute
