% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{cas.read.jmp,CAS-method}
\alias{cas.read.jmp,CAS-method}
\title{Upload a JMP File to a CAS Table}
\usage{
\S4method{cas.read.jmp}{CAS}(x, file, ..., casOut = "")
}
\arguments{
\item{x}{An instance of a CAS object that represents
a connection and CAS session.}

\item{file}{An \code{character} string that specifies
the JMP file.}

\item{\dots}{Parameters sent to the \code{table.upload} action.}

\item{casOut}{An optional \code{character} or list. If
you specify a string, then the string is used as the
in-memory table name. A list can be used to specify
properties for the in-memory table as follows:
\describe{
  \item{\code{name}}{An optional \code{character} that
    specifies the name for the in-memory table. By
    default, the name of the data frame is used.}
  \item{\code{caslib}}{An optional \code{character} that
    specifies the caslib. Specify this parameter to
    override the active caslib.}
  \item{\code{label}}{An optional \code{character} that
    specifies a descriptive label for the data.}
  \item{\code{replace}}{An optional \code{logical}. When
    set to TRUE, you can replace an existing in-memory
    table with the same name in the same caslib. The
    default value is FALSE.}
  \item{\code{promote}}{An optional \code{logical}. When
    set to TRUE, the in-memory table has global scope and
    can be available to other CAS sessions (subject to
    access controls). The default value is FALSE and
    the in-memory table has session scope so that it is
    accessible with the session that uploaded the table
    only. Session-scope tables are ideal for data analysis.
    Global-scope tables are better suited for reporting.}
  \item{\code{replication}}{An optional \code{numeric} that
    specifies the number of redundant copies of in-memory
    blocks. This parameter applies to distributed servers
    only. The default value is 1.}
 }}
}
\value{
\code{\link{CASTable}}
}
\description{
This function transfers a JMP file (.jmp) from \R
(the client) to the CAS server. The
server imports the data and \R returns a CASTable
object to reference the in-memory table in CAS.
}
\examples{
\dontrun{
spring_example <- cas.read.jmp(s, "/path/to/Spring\\\\ Example.jmp",
                               casOut = list(name = "spring_example"))
}

}
