% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.reason,CASTable-method}
\alias{cas.reason,CASTable-method}
\title{Return resulting reason for failure of last CAS action}
\usage{
\S4method{cas.reason}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
character
}
\description{
The reason is a general class of error that occurred.
Use `cas.status` for the full error message.
}
