% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.reason,list-method}
\alias{cas.reason,list-method}
\title{Return resulting reason for failure from result set}
\usage{
\S4method{cas.reason}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
character
}
\description{
The reason is a general class of error that occurred.
Use `cas.status` for the full error message.
}
