% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.retrieve,CAS-method}
\alias{cas.retrieve,CAS-method}
\title{Invoke a CAS action and return the results}
\usage{
\S4method{cas.retrieve}{CAS}(.x, .action, ..., stop.on.error = FALSE)
}
\arguments{
\item{.x}{\code{CAS} connection object.}

\item{.action}{String containing action name.}

\item{\ldots}{Action parameters.}
}
\value{

}
\description{
The \code{.action} parameter is a string containing the name of
the action. All other arguments are passed to the CAS action.
}
\details{
The results are returned in a list with the following fields:

\describe{
  \item{\code{results}}{The results of the CAS action.}
  \item{\code{messages}}{Any message printed by the action.}
  \item{\code{performance}}{Performance metrics.}
  \item{\code{disposition}}{Information about errors or warning produced
    during the action run.}
  \item{\code{events}}{Events posted by the server; this typically
    corresponds to things like caslibs and tables being updated.}
}

This method is not typically called directly (although it can be).
The generated action wrappers (e.g., \code{cas.table.columnInfo()},
\code{cas.simple.summary()}, and so on) call this method behind
the scenes and return the \code{results} field.
}
