% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.sd,CASTable-method}
\alias{cas.sd,CASTable-method}
\title{Standard Deviation}
\usage{
\S4method{cas.sd}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the standard deviation of the values for each
column in  the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.sd(ct[1:4])
cas.sd(ct$n2)
}

}
