% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.sd}
\alias{cas.sd}
\title{Standard Deviation}
\usage{
cas.sd(x, na.rm = TRUE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{An optional \code{logical}. When set to FALSE,
missing values (NA) are not removed from the analysis.
By default, missing values are ignored.}
}
\value{
A casDataFrame which includes one row for each numeric
  variable and a column that is named Std for the standard deviation.
}
\description{
Returns the standard deviation of the values for each
column in  the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.sd(ct[1:4])
cas.sd(ct$n2)
}
}
% Copyright SAS Institute
