% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.session,CAS-method}
\alias{cas.session,CAS-method}
\title{Return the session ID of the CAS connection}
\usage{
\S4method{cas.session}{CAS}(.x)
}
\arguments{
\item{.x}{\code{\link{CAS}} connection object}
}
\value{
character
}
\description{
Return the session ID of the CAS connection
}
