% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.session,CASTable-method}
\alias{cas.session,CASTable-method}
\title{Return the session ID of the CAS connection}
\usage{
\S4method{cas.session}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
character
}
\description{
Return the session ID of the CAS connection
}
