% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.severity,CASTable-method}
\alias{cas.severity,CASTable-method}
\title{Return resulting severity of last CAS action}
\usage{
\S4method{cas.severity}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
integer
}
\description{
Possible values are 0 for no errors or warnings,
1 for warnings, and 2 for errors.
}
