% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.severity,list-method}
\alias{cas.severity,list-method}
\title{Return resulting severity from result set}
\usage{
\S4method{cas.severity}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
integer
}
\description{
Possible values are 0 for no errors or warnings,
1 for warnings, and 2 for errors.
}
