% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.status,CASTable-method}
\alias{cas.status,CASTable-method}
\title{Return resulting status message for failure of last CAS action}
\usage{
\S4method{cas.status}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
character
}
\description{
Return resulting status message for failure of last CAS action
}
