% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.status,list-method}
\alias{cas.status,list-method}
\title{Return resulting status message from result set}
\usage{
\S4method{cas.status}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
character
}
\description{
Return resulting status message from result set
}
