% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.status_code,CASTable-method}
\alias{cas.status_code,CASTable-method}
\title{Return resulting status code for failure of last CAS action}
\usage{
\S4method{cas.status_code}{CASTable}(.x)
}
\arguments{
\item{.x}{\code{\link{CASTable}} object}
}
\value{
numeric
}
\description{
This status code can be helpful in tech support issues to
determine the underlying problem.
}
