% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.status_code,list-method}
\alias{cas.status_code,list-method}
\title{Return resulting status code from result set}
\usage{
\S4method{cas.status_code}{list}(.x)
}
\arguments{
\item{.x}{list}
}
\value{
numeric
}
\description{
This status code can be helpful in tech support issues to
determine the underlying problem.
}
