% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.stderr,CASTable-method}
\alias{cas.stderr,CASTable-method}
\title{Standard Error}
\usage{
\S4method{cas.stderr}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the standard error of the values for each
column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.stderr(irisct)
x["Sepal.Length"]
x[1:2]
}

}
