% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.stderr}
\alias{cas.stderr}
\title{Standard Error}
\usage{
cas.stderr(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
vector

The result is a named numeric vector. You can access
the standard error by column name or index.
}
\description{
Returns the standard error of the values for each 
column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.stderr(irisct)
x['Sepal.Length']
x[1:2]
}
}
% Copyright SAS Institute
