% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.sum,CASTable-method}
\alias{cas.sum,CASTable-method}
\title{Column Sums}
\usage{
\S4method{cas.sum}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the sum of the values for each column in
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.sum(ct[1:4])
cas.sum(ct$n2)
}

}
\seealso{
\code{colSums,CASTable-method}
}
