% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.sum}
\alias{cas.sum}
\title{Column Sums}
\usage{
cas.sum(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
casDataFrame

The result includes one row for each numeric variable
and a column that is named Sum for the summed value.
}
\description{
Returns the sum of the values for each column in 
the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.sum(ct[1:4])
cas.sum(ct$n2)
}
}
\seealso{
\code{colSums,CASTable-method}
}
% Copyright SAS Institute
