% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.terminate.CAS}
\alias{cas.terminate.CAS}
\title{End a CAS session and close the connection}
\usage{
cas.terminate.CAS(conn)
}
\arguments{
\item{conn}{The CAS connection object}
}
\description{
End a CAS session and close the connection
}
\seealso{
\code{\link{cas.close}}
}
