% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.terminate}
\alias{cas.terminate}
\title{End a CAS session and close the connection}
\usage{
cas.terminate(conn)
}
\arguments{
\item{CAS}{The CAS connection object}
}
\description{
End a CAS session and close the connection
}
