% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.tvalue}
\alias{cas.tvalue}
\title{T-Statistics for Hypothesis Testing}
\usage{
cas.tvalue(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
vector

The result is a named numeric vector. You can access
the t-statistic by column name or index.
}
\description{
Returns the t-statistic for the
values for each column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.tvalue(irisct)
x['Sepal.Length']
x[1:2]
}
}
% Copyright SAS Institute
