% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.upload.file,CAS-method}
\alias{cas.upload.file,CAS-method}
\title{Upload a data file to a CAS table}
\usage{
\S4method{cas.upload.file}{CAS}(.x, .file, ...)
}
\arguments{
\item{\dots}{Optional parameters that are passed to the table.loadtable action.}

\item{x}{The \code{\link{CAS}} connection object}

\item{file}{The filename to upload}
}
\value{
\code{\link{CASTable}} object which references a new CAS table
  on the server that contains the data from the given \code{data.frame}.
}
\description{
Upload a data file to a CAS table
}
\seealso{
\code{\link{cas.upload.frame}}
}
