% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.upload}
\alias{cas.upload}
\title{Upload a data.frame or file to a CAS table}
\usage{
cas.upload(conn, ...)
}
\arguments{
\item{\dots}{Optional parameters that are passed to the table.loadtable action.}

\item{CAS}{The CAS connection object}

\item{data.frame/character}{The data.frame, filename, or URL to upload.}
}
\value{
\code{list}
}
\description{
Upload a data.frame or file to a CAS table
}
