% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.uss}
\alias{cas.uss}
\title{Uncorrected Sum of Squares}
\usage{
cas.uss(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
vector

The result is a named numeric vector. You can access
the uncorrected sum of squares by column name or index.
}
\description{
Returns the uncorrected sum of squares of the 
values for each column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
x <- cas.uss(irisct)
x['Sepal.Length']
x[1:2]
}
}
% Copyright SAS Institute
