% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.var,CASTable-method}
\alias{cas.var,CASTable-method}
\title{Variance}
\usage{
\S4method{cas.var}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Returns the variance of the values for each
column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.var(ct[1:4])
cas.var(ct$n2)
}

}
