% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cas.var}
\alias{cas.var}
\title{Variance}
\usage{
cas.var(CASTable)
}
\arguments{
\item{x}{CASTable.}
}
\value{
casDataFrame

The result includes one row for each numeric variable
and a column that is named Var for the variance.
}
\description{
Returns the variance of the values for each 
column in the input table.
}
\details{
This function operates on numeric columns only.
}
\examples{
\dontrun{
cas.var(ct[1:4])
cas.var(ct$n2)
}
}
% Copyright SAS Institute
