% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{cas.write.table}
\alias{cas.write.table}
\title{Write a CAS Table to a Table}
\usage{
cas.write.table(CASTable, file = "", append = FALSE, quote = TRUE,
  sep = " ", eol = "\\n", na = "NA", dec = ".", row.names = TRUE,
  col.names = TRUE, qmethod = c("escape", "double"), fileEncoding = "")
}
\arguments{
\item{CASTable}{The instance of the CASTable to save as
as a file.}

\item{file}{An \code{character} string that 
specifies the filename for the file. If you do not
specify the file, then the table is printed to the 
terminal.}

\item{append}{An optional \code{logical} value. This value
is passed to \code{write.table}.}

\item{quote}{An optional \code{logical} value or numeric
vector. This value is passed to \code{write.table}.}

\item{sep}{An optional \code{character} that is used to separate
values in the file. This value is passed to \code{write.table}.}

\item{eol}{An optional \code{character} string that is used as the
end-of-line character or characters. This value is passed to 
\code{write.table}.}

\item{na}{An optional \code{character} string to represent
missing values. This value is passed to \code{write.table}.}

\item{dec}{An optional \code{character} to represent the decimal
separator. This value is passed to \code{write.table}.}

\item{row.names}{An optional \code{logical} value or a 
\code{character} vector of row names. This value is passed to
\code{write.table}.}

\item{col.names}{An optional \code{logical} value or a
\code{character} vector of column names. This value is passed
to \code{write.table}.}

\item{qmethod}{An optional \code{chracter} string that describes
how to write embedded quotation marks. This value is passed
to \code{write.table}.}

\item{fileEncoding}{An optional \code{character} string that
specifies the encoding to use for writing the file. This
value is passed to \code{write.table}.}
}
\description{
This function downloads an in-memory table from the
CAS server and saves it as a file that is accesible
to \R (the client). This function is a convenience
wrapper for the \R \code{write.table} function.
}
\examples{
\dontrun{
# 
cas.write.table(myCasTable, file="/path/to/data_out.txt", na="")
}
}
\seealso{
Other functions for saving in-memory data: \code{\link{cas.saveRDS}},
  \code{\link{cas.write.csv}}, \code{\link{cas.write.xlsx}}
}
% Copyright SAS Institute
