% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASd.as.Date}
\alias{CASd.as.Date}
\title{Convert a CAS date to Date}
\usage{
CASd.as.Date(cdt)
}
\arguments{
\item{cdt}{Numeric representing a CAS date}
}
\value{
Date
}
\description{
Convert a CAS date to Date
}
