% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASd.as.SASd}
\alias{CASd.as.SASd}
\title{Convert a CAS date to a SAS date}
\usage{
CASd.as.SASd(cdt)
}
\arguments{
\item{cdt}{Numeric representing a CAS date}
}
\value{
Numeric representing a SAS date
}
\description{
Convert a CAS date to a SAS date
}
