% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_df.R
\docType{class}
\name{casDataFrame-class}
\alias{casDataFrame-class}
\alias{casDataFrame}
\title{CAS Data Frame Class}
\arguments{
\item{\ldots}{Arguments used to construct a data.frame.}

\item{name}{Optional \code{character} string that specifies
CAS metadata for a table name.}

\item{label}{Optional \code{character} string that specifies
CAS metadata for a table label.}

\item{title}{Optional \code{character} string that specifies
CAS metadata for a table title.}

\item{attrs}{Optional \code{list} of key-value pairs the specify
user-defined CAS metadata.}

\item{col.labels}{Optional \code{character} string that specifies CAS
metadata for column labels.}

\item{col.formats}{Optional \code{character} string that specifies CAS
metadata for column formats.}

\item{col.attrs}{Optional \code{list} that specifies CAS metadata for
column attributes.}

\item{col.sizes}{Optional \code{list} that specifies CAS metadata for
the number of bytes in the widest row.}

\item{col.types}{Optional \code{character} that specifies CAS metadata
for the column data types.}

\item{col.widths}{Optional \code{numeric} that specifies CAS metadata
for column widths.}
}
\description{
A casDataFrame is a superset of data.frame. The class is used primarily
by functions within the package to store tabular data from CAS action
results and to associate CAS metadata with the tabular data.
}
\details{
A casDataFrame instance is a data object in \R (the client).
}
\section{Slots}{

\describe{
\item{\code{name}}{\code{character} string that contains the table name.}

\item{\code{label}}{\code{character} string that contains the table label.}

\item{\code{title}}{\code{character} string that contains the table title.}

\item{\code{attrs}}{\code{list} of key-value pairs the specify user-defined
CAS metadata.}

\item{\code{col.labels}}{\code{character} vector that contains the column labels.}

\item{\code{col.formats}}{\code{character} vector that contains the column formats.}

\item{\code{col.attrs}}{\code{list} that contains the column attributes.}

\item{\code{col.sizes}}{\code{list} that contains the byte-width of the columns.}

\item{\code{col.types}}{\code{character} that contains the column data types.}

\item{\code{col.widths}}{\code{numeric} that contains the column widths.}

\item{\code{df}}{The \code{data.frame} which contains the underlying data.}
}}

