% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASdt.as.POSIXct}
\alias{CASdt.as.POSIXct}
\title{Convert CAS timestamp to POSIXct}
\usage{
CASdt.as.POSIXct(cts)
}
\arguments{
\item{cts}{Numeric representing a CAS timestamp}
}
\value{
POSIXct object
}
\description{
Convert CAS timestamp to POSIXct
}
