% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASdt.as.POSIXlt}
\alias{CASdt.as.POSIXlt}
\title{Convert CAS timestamp to POSIXlt}
\usage{
CASdt.as.POSIXlt(cts)
}
\arguments{
\item{cts}{Numeric representing a CAS timestamp}
}
\value{
POSIXlt object
}
\description{
Convert CAS timestamp to POSIXlt
}
