% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASdt.as.SASdt}
\alias{CASdt.as.SASdt}
\title{Convert a CAS timestamp to a SAS timestamp}
\usage{
CASdt.as.SASdt(cts)
}
\arguments{
\item{cts}{Numeric representing a CAS timestamp}
}
\value{
Numeric representing a SAS timestamp
}
\description{
Convert a CAS timestamp to a SAS timestamp
}
