% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\docType{class}
\name{CASEventWatcher-class}
\alias{CASEventWatcher-class}
\alias{CASEventWatcher}
\title{Event watcher for multiple CAS connections}
\arguments{
\item{connections}{A list of \code{CAS} connections.}

\item{\ldots}{Options for the event watcher including the following:
\describe{
  \item{timeout}{Integer indicating the connection timeout in seconds.}
  \item{datamsghandler}{A \code{CASDataMsgHandler} class used to handle
    data upload requests when calling the \code{table.addtable} action.}
}}
}
\description{
Event watchers monitor one or more CAS connections for responses coming
from the server. When a connection gets a response, the event watcher
returns the connection where the event occurred. This process is
handled in the \code{getnext()} function.
}
\section{Fields}{

\describe{
\item{\code{sw_watcher}}{The watcher object in the C extension.}

\item{\code{connections}}{A list of \code{CAS} connections.}

\item{\code{datamsghandler}}{A data message handler class used for actions
that require data uploads.}
}}


\keyword{internal}
