% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{casFormula}
\alias{casFormula}
\title{CAS Function to extract information from an \R formula object}
\usage{
casFormula(f, ct)
}
\arguments{
\item{f1}{a formula object}
}
\value{
list with target and inputs as members
}
\description{
CAS Function to extract information from an \R formula object
}
\keyword{internal}
% Copyright SAS Institute
