% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{$,CASTable-method}
\alias{$,CASTable-method}
\title{Access CASTable column by name}
\usage{
\S4method{$}{CASTable}(x, name)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{name}{String containing column name.}
}
\value{
\code{\link{CASTable}} with just specified column visible.
}
\description{
Access CASTable column by name
}
