% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{$<-,CASTable-method}
\alias{$<-,CASTable-method}
\alias{$,CASTable-method}
\title{Assign data to column by name}
\usage{
\S4method{$}{CASTable}(x, name) <- value
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{name}{String containing column name.}

\item{value}{Value to assign.}
}
\value{
\code{\link{CASTable}} object.
}
\description{
Assign data to column by name
}
