% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\docType{class}
\name{CASRequest-class}
\alias{CASRequest-class}
\alias{CASRequest}
\title{Object containing a CAS server request}
\arguments{
\item{sw_request}{The C extension request object.}
}
\description{
Object containing a CAS server request
}

\keyword{internal}
