% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\docType{class}
\name{CASResponse-class}
\alias{CASResponse-class}
\alias{CASResponse}
\title{Object containing a response from the CAS server}
\arguments{
\item{sw_response}{The C extension response object.}
}
\description{
This object is typically used behind-the-scenes in the CAS action
wrappers, but is not surfaced. The one use-case where it may be
encountered is when using the \code{CAS$invoke()} method and 
retrieving the responses from the server manually.
}
\section{Fields}{

\describe{
\item{\code{disposition}}{List containing information about the final
status of the CAS action run.}

\item{\code{performance}}{List containing performance metrics of the
CAS action.}

\item{\code{messages}}{List of messages produced by the CAS action.}
}}


\keyword{internal}
