% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{casRetrieve}
\alias{casRetrieve}
\title{Utility function that is run from several exported functions}
\usage{
casRetrieve(conn, ...)
}
\description{
Utility function that is run from several exported functions
}
\keyword{internal}
