% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASt.as.POSIXct}
\alias{CASt.as.POSIXct}
\title{Convert a CAS time to POSIXct}
\usage{
CASt.as.POSIXct(ctm)
}
\arguments{
\item{ctm}{Numeric representing a CAS time}
}
\value{
POSIXct object
}
\description{
Convert a CAS time to POSIXct
}
