% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASt.as.POSIXlt}
\alias{CASt.as.POSIXlt}
\title{Convert a CAS time to POSIXlt}
\usage{
CASt.as.POSIXlt(ctm)
}
\arguments{
\item{ctm}{Numeric representing a CAS time}
}
\value{
POSIXlt object
}
\description{
Convert a CAS time to POSIXlt
}
