% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{CASt.as.SASt}
\alias{CASt.as.SASt}
\title{Convert a CAS time to a SAS time}
\usage{
CASt.as.SASt(ctm)
}
\arguments{
\item{ctm}{Numeric representing a CAS time}
}
\value{
Numeric representing a SAS time
}
\description{
Convert a CAS time to a SAS time
}
