% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{CASTable-arithmetic}
\alias{CASTable-arithmetic}
\alias{.cas_arith}
\alias{-,CASTable,CASTable-method}
\alias{-,CASTable,ANY-method}
\alias{-,ANY,CASTable-method}
\alias{+,CASTable,CASTable-method}
\alias{+,CASTable,ANY-method}
\alias{+,ANY,CASTable-method}
\alias{/,CASTable,CASTable-method}
\alias{/,CASTable,ANY-method}
\alias{/,ANY,CASTable-method}
\alias{*,CASTable,CASTable-method}
\alias{*,CASTable,ANY-method}
\alias{*,ANY,CASTable-method}
\alias{^,CASTable,CASTable-method}
\alias{^,CASTable,ANY-method}
\alias{^,ANY,CASTable-method}
\alias{\%\%,CASTable,CASTable-method}
\alias{\%\%,CASTable,ANY-method}
\alias{\%\%,ANY,CASTable-method}
\alias{\%/\%,CASTable,CASTable-method}
\alias{\%/\%,CASTable,ANY-method}
\alias{\%/\%,ANY,CASTable-method}
\title{CASTable arithmetic operators}
\usage{
.cas_arith(e1, op, e2)

\S4method{-}{CASTable,CASTable}(e1, e2)

\S4method{-}{CASTable,ANY}(e1, e2)

\S4method{-}{ANY,CASTable}(e1, e2)

\S4method{+}{CASTable,CASTable}(e1, e2)

\S4method{+}{CASTable,ANY}(e1, e2)

\S4method{+}{ANY,CASTable}(e1, e2)

\S4method{/}{CASTable,CASTable}(e1, e2)

\S4method{/}{CASTable,ANY}(e1, e2)

\S4method{/}{ANY,CASTable}(e1, e2)

\S4method{*}{CASTable,CASTable}(e1, e2)

\S4method{*}{CASTable,ANY}(e1, e2)

\S4method{*}{ANY,CASTable}(e1, e2)

\S4method{^}{CASTable,CASTable}(e1, e2)

\S4method{^}{CASTable,ANY}(e1, e2)

\S4method{^}{ANY,CASTable}(e1, e2)

\S4method{\%\%}{CASTable,CASTable}(e1, e2)

\S4method{\%\%}{CASTable,ANY}(e1, e2)

\S4method{\%\%}{ANY,CASTable}(e1, e2)

\S4method{\%/\%}{CASTable,CASTable}(e1, e2)

\S4method{\%/\%}{CASTable,ANY}(e1, e2)

\S4method{\%/\%}{ANY,CASTable}(e1, e2)
}
\description{
CASTable arithmetic operators
}
\keyword{internal}
