% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{CASTable-comparison}
\alias{CASTable-comparison}
\alias{.cas_compare}
\alias{>,CASTable,CASTable-method}
\alias{>,CASTable,ANY-method}
\alias{>,ANY,CASTable-method}
\alias{<,CASTable,CASTable-method}
\alias{<,CASTable,ANY-method}
\alias{<,ANY,CASTable-method}
\alias{>=,CASTable,CASTable-method}
\alias{>=,CASTable,ANY-method}
\alias{>=,ANY,CASTable-method}
\alias{<=,CASTable,CASTable-method}
\alias{<=,CASTable,ANY-method}
\alias{<=,ANY,CASTable-method}
\alias{==,CASTable,CASTable-method}
\alias{==,CASTable,ANY-method}
\alias{==,ANY,CASTable-method}
\alias{!=,CASTable,CASTable-method}
\alias{!=,CASTable,ANY-method}
\alias{!=,ANY,CASTable-method}
\title{CASTable comparison operators}
\usage{
.cas_compare(e1, op, e2)

\S4method{>}{CASTable,CASTable}(e1, e2)

\S4method{>}{CASTable,ANY}(e1, e2)

\S4method{>}{ANY,CASTable}(e1, e2)

\S4method{<}{CASTable,CASTable}(e1, e2)

\S4method{<}{CASTable,ANY}(e1, e2)

\S4method{<}{ANY,CASTable}(e1, e2)

\S4method{>=}{CASTable,CASTable}(e1, e2)

\S4method{>=}{CASTable,ANY}(e1, e2)

\S4method{>=}{ANY,CASTable}(e1, e2)

\S4method{<=}{CASTable,CASTable}(e1, e2)

\S4method{<=}{CASTable,ANY}(e1, e2)

\S4method{<=}{ANY,CASTable}(e1, e2)

\S4method{==}{CASTable,CASTable}(e1, e2)

\S4method{==}{CASTable,ANY}(e1, e2)

\S4method{==}{ANY,CASTable}(e1, e2)

\S4method{!=}{CASTable,CASTable}(e1, e2)

\S4method{!=}{CASTable,ANY}(e1, e2)

\S4method{!=}{ANY,CASTable}(e1, e2)
}
\description{
CASTable comparison operators
}
\keyword{internal}
