% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R, R/Ops.R
\docType{methods}
\name{[,CASTable-method}
\alias{[,CASTable-method}
\alias{[<-,CASTable-method}
\alias{[[,CASTable-method}
\alias{$,CASTable-method}
\alias{$<-,CASTable-method}
\alias{.cas.arith}
\alias{-,CASTable,CASTable-method}
\alias{-,CASTable,ANY-method}
\alias{-,ANY,CASTable-method}
\alias{+,CASTable,CASTable-method}
\alias{+,CASTable,ANY-method}
\alias{+,ANY,CASTable-method}
\alias{/,CASTable,CASTable-method}
\alias{/,CASTable,ANY-method}
\alias{/,ANY,CASTable-method}
\alias{*,CASTable,CASTable-method}
\alias{*,CASTable,ANY-method}
\alias{*,ANY,CASTable-method}
\alias{^,CASTable,CASTable-method}
\alias{^,CASTable,ANY-method}
\alias{^,ANY,CASTable-method}
\alias{\%\%,CASTable,CASTable-method}
\alias{\%\%,CASTable,ANY-method}
\alias{\%\%,ANY,CASTable-method}
\alias{\%/\%,CASTable,CASTable-method}
\alias{\%/\%,CASTable,ANY-method}
\alias{\%/\%,ANY,CASTable-method}
\alias{.cas.compare}
\alias{>,CASTable,CASTable-method}
\alias{>,CASTable,ANY-method}
\alias{>,ANY,CASTable-method}
\alias{<,CASTable,CASTable-method}
\alias{<,CASTable,ANY-method}
\alias{<,ANY,CASTable-method}
\alias{>=,CASTable,CASTable-method}
\alias{>=,CASTable,ANY-method}
\alias{>=,ANY,CASTable-method}
\alias{<=,CASTable,CASTable-method}
\alias{<=,CASTable,ANY-method}
\alias{<=,ANY,CASTable-method}
\alias{==,CASTable,CASTable-method}
\alias{==,CASTable,ANY-method}
\alias{==,ANY,CASTable-method}
\alias{!=,CASTable,CASTable-method}
\alias{!=,CASTable,ANY-method}
\alias{!=,ANY,CASTable-method}
\alias{.cas.logic}
\alias{&,CASTable,CASTable-method}
\alias{&,CASTable,ANY-method}
\alias{&,ANY,CASTable-method}
\alias{|,CASTable,CASTable-method}
\alias{|,CASTable,ANY-method}
\alias{|,ANY,CASTable-method}
\alias{!,CASTable-method}
\title{Extract Columns from a CAS Table}
\usage{
\S4method{[}{CASTable}(x, i, j, ..., drop = TRUE)

\S4method{[}{CASTable}(x, i, j, ...) <- value

\S4method{[[}{CASTable}(x, i)

\S4method{$}{CASTable}(x, name)

\S4method{$}{CASTable}(x, name) <- value

.cas.arith(e1, op, e2)

\S4method{-}{CASTable,CASTable}(e1, e2)

\S4method{-}{CASTable,ANY}(e1, e2)

\S4method{-}{ANY,CASTable}(e1, e2)

\S4method{+}{CASTable,CASTable}(e1, e2)

\S4method{+}{CASTable,ANY}(e1, e2)

\S4method{+}{ANY,CASTable}(e1, e2)

\S4method{/}{CASTable,CASTable}(e1, e2)

\S4method{/}{CASTable,ANY}(e1, e2)

\S4method{/}{ANY,CASTable}(e1, e2)

\S4method{*}{CASTable,CASTable}(e1, e2)

\S4method{*}{CASTable,ANY}(e1, e2)

\S4method{*}{ANY,CASTable}(e1, e2)

\S4method{^}{CASTable,CASTable}(e1, e2)

\S4method{^}{CASTable,ANY}(e1, e2)

\S4method{^}{ANY,CASTable}(e1, e2)

\S4method{\%\%}{CASTable,CASTable}(e1, e2)

\S4method{\%\%}{CASTable,ANY}(e1, e2)

\S4method{\%\%}{ANY,CASTable}(e1, e2)

\S4method{\%/\%}{CASTable,CASTable}(e1, e2)

\S4method{\%/\%}{CASTable,ANY}(e1, e2)

\S4method{\%/\%}{ANY,CASTable}(e1, e2)

.cas.compare(e1, op, e2)

\S4method{>}{CASTable,CASTable}(e1, e2)

\S4method{>}{CASTable,ANY}(e1, e2)

\S4method{>}{ANY,CASTable}(e1, e2)

\S4method{<}{CASTable,CASTable}(e1, e2)

\S4method{<}{CASTable,ANY}(e1, e2)

\S4method{<}{ANY,CASTable}(e1, e2)

\S4method{>=}{CASTable,CASTable}(e1, e2)

\S4method{>=}{CASTable,ANY}(e1, e2)

\S4method{>=}{ANY,CASTable}(e1, e2)

\S4method{<=}{CASTable,CASTable}(e1, e2)

\S4method{<=}{CASTable,ANY}(e1, e2)

\S4method{<=}{ANY,CASTable}(e1, e2)

\S4method{==}{CASTable,CASTable}(e1, e2)

\S4method{==}{CASTable,ANY}(e1, e2)

\S4method{==}{ANY,CASTable}(e1, e2)

\S4method{!=}{CASTable,CASTable}(e1, e2)

\S4method{!=}{CASTable,ANY}(e1, e2)

\S4method{!=}{ANY,CASTable}(e1, e2)

.cas.logic(e1, op, e2)

\S4method{&}{CASTable,CASTable}(e1, e2)

\S4method{&}{CASTable,ANY}(e1, e2)

\S4method{&}{ANY,CASTable}(e1, e2)

\S4method{|}{CASTable,CASTable}(e1, e2)

\S4method{|}{CASTable,ANY}(e1, e2)

\S4method{|}{ANY,CASTable}(e1, e2)

\S4method{!}{CASTable}(x)
}
\description{
Extract Columns from a CAS Table
}
% Copyright SAS Institute
