% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{CASTable-logical}
\alias{CASTable-logical}
\alias{.cas_logic}
\alias{&,CASTable,CASTable-method}
\alias{&,CASTable,ANY-method}
\alias{&,ANY,CASTable-method}
\alias{|,CASTable,CASTable-method}
\alias{|,CASTable,ANY-method}
\alias{|,ANY,CASTable-method}
\alias{!,CASTable-method}
\title{CASTable logical operators}
\usage{
.cas_logic(e1, op, e2)

\S4method{&}{CASTable,CASTable}(e1, e2)

\S4method{&}{CASTable,ANY}(e1, e2)

\S4method{&}{ANY,CASTable}(e1, e2)

\S4method{|}{CASTable,CASTable}(e1, e2)

\S4method{|}{CASTable,ANY}(e1, e2)

\S4method{|}{ANY,CASTable}(e1, e2)

\S4method{!}{CASTable}(x)
}
\description{
CASTable logical operators
}
\keyword{internal}
