% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{cbind,CASTable-method}
\alias{cbind,CASTable-method}
\title{Combine CAS Tables by Columns}
\usage{
\S4method{cbind}{CASTable}(..., deparse.level = 1)
}
\arguments{
\item{...}{Arguments that are passed to method arguments.}

\item{deparse.level}{See the help for base::rbind.}
}
\description{
Combine CAS Tables by Columns
}
