% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{cbind2.casTable}
\alias{cbind2.casTable}
\alias{cbind2,CASTable,ANY-method}
\title{Combine CAS Tables by Columns}
\usage{
cbind2.casTable(x, y, ...)

\S4method{cbind2}{CASTable,ANY}(x, y, ...)
}
\arguments{
\item{\dots}{Arguments that are passed to method arguments.}

\item{deparse.level}{See the help for base::cbind.}
}
\value{
\code{\link{CASTable}}
}
\description{
This is the implementation of \code{cbind} for in-memory tables.
}
\examples{
\dontrun{
cbind(ct1, ct2)
cbind(ct1[1:3], ct2$X1)
}
}
% Copyright SAS Institute
