% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{colMeans,CASTable-method}
\alias{colMeans,CASTable-method}
\title{Column Means}
\usage{
\S4method{colMeans}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Column Means
}
\examples{
\dontrun{
colMeans(ct[1:4])
colMeans(ct$X1)
}
}
\seealso{
\code{cas.mean}
}
