% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{colnames,CASTable-method}
\alias{colnames,CASTable-method}
\title{Column Names in a CAS Table}
\usage{
\S4method{colnames}{CASTable}(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
vector
}
\description{
Returns the column names from the in-memory table
that is referenced by the \code{\link{CASTable}} object.
}
\section{Note}{

You cannot use this function to set the column names.
}

\examples{
\dontrun{
colnames(ct1)
colnames(ct[1:4])
}

}
\seealso{
\code{names,CASTable-method}
}
