% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyAllExcept.R
\name{copyAllExcept}
\alias{copyAllExcept}
\title{Copy all files from one folder to another folder except some files}
\usage{
copyAllExcept(fromDir, toDir, exceptFiles)
}
\arguments{
\item{fromDir}{path of the directory contains files being copied}

\item{toDir}{path of the destination directory where the files are copied to}

\item{exceptFiles}{character vector of files that do not need to copy}
}
\value{
no return, copied files will be saved in toDir
}
\description{
Copy all files from one folder to another folder except some files
}
\examples{
\donttest{
# Create a directory and populate with TxtInOut of SWAT+
extracExampleData(exampleData,"swatPlusTxtInOut", tempdir())
fromDir <- file.path(tempdir(), "swatPlusTxtInOut")

# Create a directory to copy to this directory
dir.create(file.path(tempdir(), "toDir"), showWarnings = FALSE)
toDir <- file.path(tempdir(), "toDir")

# Except these files
exceptFiles <- exampleData$swatPlusFiles[1:73]

# Not copy all files from fromDir to toDir except exceptFiles
copyAllExcept(fromDir, toDir, exceptFiles)
}
}
