% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\docType{methods}
\name{cor,CASTable-method}
\alias{cor,CASTable-method}
\title{Correlation}
\usage{
\S4method{cor}{CASTable}(x, y = NULL, use = "everything",
  method = c("pearson"))
}
\arguments{
\item{x}{CASTable.}
}
\value{
matrix
}
\description{
Unlike the \code{cor} function for data frames, 
this function does not support specifying the
method as "kendall" or "spearman." The results
for method "pearson" are returned.
}
\examples{
\dontrun{
cor(ct1)
cor(i2[1:4])
cor(i2[c('col1', 'col2', 'col3')])
cor(i2$col1, i2$col2)
}
}
% Copyright SAS Institute
