% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDirCopyUnchangeFile.R
\name{createDirCopyUnchangeFile}
\alias{createDirCopyUnchangeFile}
\title{Create TxtInOut directories and copy unchanged files}
\usage{
createDirCopyUnchangeFile(
  workingDirectory,
  ncores,
  TxtInOut,
  exceptFiles,
  swatExe,
  firstRun
)
}
\arguments{
\item{workingDirectory}{path to the working directory where all files created
by SWAT or SWAT+ will be saved}

\item{ncores}{number of threads for parallel execution}

\item{TxtInOut}{path to original TxtInOut folder}

\item{exceptFiles}{character vector of files that do not need to copy}

\item{swatExe}{SWAT/SWAT+ executable files}

\item{firstRun}{logical (set to TRUE if this is the first iteration/run),
in the first iteration, all files need to be copied to the new TxtInOut folder
for SWAT/SWAT+ run, in the subsequent iterations, only files that content
parameters (that need to be changed) are updated.}
}
\value{
No return
}
\description{
Create TxtInOut directories and copy unchanged files
}
\examples{
\donttest{
# Create a directory and populate with TxtInOut of SWAT
extracExampleData(exampleData,"swatTxtInOut", tempdir())
TxtInOut <- file.path(tempdir(), "swatTxtInOut")

# Create a directory to copy to this directory
dir.create(file.path(tempdir(), "workingDirectory"), showWarnings = FALSE)
workingDirectory <- file.path(tempdir(), "workingDirectory")

# Except these files
exceptFiles <- exampleData$swatFiles[1:280]

# Lets assume that this is the SWAT exe file
swatExe <- tempfile(pattern = "swatexe", tmpdir = tempdir(), fileext = ".exxxe")
file.create(swatExe)

createDirCopyUnchangeFile(workingDirectory, 2, TxtInOut, exceptFiles, swatExe, TRUE)
}

}
