% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{Date.as.CASd}
\alias{Date.as.CASd}
\title{Convert a Date to a CAS date}
\usage{
Date.as.CASd(rdt)
}
\arguments{
\item{rdt}{Date object}
}
\value{
Numeric representing a CAS date
}
\description{
Convert a Date to a CAS date
}
