% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{Date.as.SASd}
\alias{Date.as.SASd}
\title{Convert a Date to a SAS date}
\usage{
Date.as.SASd(rdt)
}
\arguments{
\item{rdt}{Date object}
}
\value{
Numeric representing a SAS date
}
\description{
Convert a Date to a SAS date
}
