% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.R
\name{dds}
\alias{dds}
\title{Parameter optimization using Dynamically Dimensioned Search (DDS)}
\usage{
dds(para, ncores, iter, nIters, r, parallelMode)
}
\arguments{
\item{para}{data frame of parameter, its mean, max and actual parameter values}

\item{ncores}{number of cores for parallel run}

\item{iter}{current iteration number}

\item{nIters}{total number of iterations}

\item{r}{TOO}

\item{parallelMode}{TODO}
}
\description{
Parameter optimization using Dynamically Dimensioned Search (DDS)
}
\examples{
\donttest{
para <- data.frame(min = c(0.1, 0.2, 0.3, 0.4),
                  max = c(1,2,3,4),
                  paraValue = c(0.3, 0.7, 1.3, 2.4))
ncores <- 3
iter <- 3
nIters <- 20
r <- 0.2
parallelMode <- 0
}

}
