% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R
\docType{methods}
\name{dim,CASTable-method}
\alias{dim,CASTable-method}
\title{Dimensions of a CAS Table}
\usage{
\S4method{dim}{CASTable}(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
vector
}
\description{
Returns the number of rows and columns for the in-memory
table that is referenced by the \code{\link{CASTable}} object.
}
\examples{
\dontrun{
dim(ct1)
}
}
% Copyright SAS Institute
