% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{dimnames,CASTable-method}
\alias{dimnames,CASTable-method}
\title{Dimension Names of a CAS Table}
\usage{
\S4method{dimnames}{CASTable}(x)
}
\arguments{
\item{x}{A \code{\link{CASTable}} object.}
}
\value{
list
}
\description{
Dimension Names of a CAS Table
}
\examples{
\dontrun{
dimnames(ct1)
dimnames(ct[2:4])
}

}
