% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.binary_enabled}
\alias{.binary_enabled}
\title{Determine if binary communication with a CAS server is possible}
\usage{
.binary_enabled()
}
\description{
Determine if binary communication with a CAS server is possible
}
\keyword{internal}
