% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_rest.R
\name{.camel2underscore}
\alias{.camel2underscore}
\title{Convert camel-case name to underscore-delimited}
\usage{
.camel2underscore(str)
}
\description{
Convert camel-case name to underscore-delimited
}
\keyword{internal}
