% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.cas2r}
\alias{.cas2r}
\title{Convert a CASValue to an R object}
\usage{
.cas2r(sw_value)
}
\arguments{
\item{sw_value}{The \code{CASValue} to convert}
}
\value{
An R object corresponding to the given value.
}
\description{
Convert a CASValue to an R object
}
\keyword{internal}
