% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.cas_formula}
\alias{.cas_formula}
\title{CAS Function to extract information from an \R formula object}
\usage{
.cas_formula(f, ct)
}
\arguments{
\item{f}{A formula object}

\item{ct}{CASTable object}
}
\value{
list with target and inputs as members
}
\description{
CAS Function to extract information from an \R formula object
}
\keyword{internal}
