% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{.casout}
\alias{.casout}
\title{Construct a casOut parameter based on the given parameters}
\usage{
.casout(out, file)
}
\arguments{
\item{out}{List containing given casOut parameter}

\item{file}{Input data file name}
}
\value{
List contaning new casOut parameter
}
\description{
Construct a casOut parameter based on the given parameters
}
\keyword{internal}
