% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.casvaluelist2r}
\alias{.casvaluelist2r}
\title{Convert a CAS value list to an R list}
\usage{
.casvaluelist2r(sw_values, len)
}
\arguments{
\item{sw_values}{A \code{CASValueList} object.}

\item{len}{The length of the CAS value list}
}
\description{
Convert a CAS value list to an R list
}
\keyword{internal}
