% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.check_for_cas_errors}
\alias{.check_for_cas_errors}
\title{Check if CAS submission had errors}
\usage{
.check_for_cas_errors(result, stop.on.error = TRUE)
}
\arguments{
\item{result}{The result to check}

\item{stop.on.error}{Should an error in the CAS log raise and exception?}
}
\value{
logical
}
\description{
The function scans the result messages from CAS to see if they contain
and ERROR string. There is an optional argument stop.on.error which
defaults to TRUE. which will stop processing, if set to FALSE a warning
will be displayed and processing will continue.
}
\keyword{internal}
