% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.combine_coverage_files}
\alias{.combine_coverage_files}
\title{Combine multiple coverage report files into one}
\usage{
.combine_coverage_files(..., file = "coverage.rds", ignore.missing = TRUE)
}
\arguments{
\item{...}{one or more filenames containing covr::coverage objects}
}
\description{
Combine multiple coverage report files into one
}
\keyword{internal}
