% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.combine_reports}
\alias{.combine_reports}
\title{Combine multiple coverage reports into one}
\usage{
.combine_reports(...)
}
\arguments{
\item{...}{one or more covr::coverage objects}
}
\value{
covr::coverage
}
\description{
Combine multiple coverage reports into one
}
\keyword{internal}
