% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.detect_protocol}
\alias{.detect_protocol}
\title{Attempt to detect the protocol of the CAS server}
\usage{
.detect_protocol(hostname, port, protocol)
}
\arguments{
\item{hostname}{The hostname of the CAS server.}

\item{port}{The port of the CAS server.}

\item{protocol}{The protocol of the CAS server; can be \code{"auto"}.}
}
\value{
String containing the detected protocol of the server.
}
\description{
Attempt to detect the protocol of the CAS server
}
\keyword{internal}
