% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.error_check}
\alias{.error_check}
\title{Check a C extension object for error messages}
\usage{
.error_check(x)
}
\arguments{
\item{x}{C extension object.}
}
\description{
If the object does have an error message attached, the \code{stop()}
function is called with that message.
}
\keyword{internal}
