% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_rest.R
\name{.expand_params}
\alias{.expand_params}
\title{Convert CASTable objects to list parameters}
\usage{
.expand_params(params)
}
\description{
Convert CASTable objects to list parameters
}
\keyword{internal}
