% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.expand_url}
\alias{.expand_url}
\title{Expand URL macro}
\usage{
.expand_url(hostname)
}
\arguments{
\item{hostname}{String containing a URL with possible expansions.
Portions of the string contained within square brackets can
contain comma-separated lists of permutations.}
}
\value{
List of resulting hostnames.
}
\description{
Expand URL macro
}
\keyword{internal}
