% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{.gen_functions}
\alias{.gen_functions}
\title{Generate action function wrappers using actionset reflection information}
\usage{
.gen_functions(x, actionset)
}
\description{
Generate action function wrappers using actionset reflection information
}
\keyword{internal}
