% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.gen_table_param}
\alias{.gen_table_param}
\title{Generate Table Parameters}
\usage{
.gen_table_param(ct)
}
\arguments{
\item{ct}{CASTable object.}
}
\value{
List containing CASTable parameters
}
\description{
Generate Table Parameters
}
\keyword{internal}
