% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.get_connection_info}
\alias{.get_connection_info}
\title{Resolve the connection information from given parameters and env variables}
\usage{
.get_connection_info(hostname, port, username, password, protocol, path)
}
\arguments{
\item{hostname}{The CAS server hostname.}

\item{port}{The CAS server port number.}

\item{username}{CAS username.}

\item{password}{CAS password.}

\item{protocol}{CAS server protocol. This should be one of the following:
\code{cas}, \code{http}, or \code{https}.}

\item{path}{The base path of the CAS server. This is only valid
when using HTTP(S) connections.}
}
\value{
List containing all connection parmeters resolved.
}
\description{
Resolve the connection information from given parameters and env variables
}
\keyword{internal}
