% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{.list_action_params}
\alias{.list_action_params}
\title{List CAS action parameters by name}
\usage{
.list_action_params(x, action, display = TRUE)
}
\arguments{
\item{x}{An instance of a CAS object that represents
a connection and CAS session.}

\item{action}{A \code{string} value that specifies
the action name. You can specify the following forms:
\itemize{
  \item \emph{action-set.action-name}
  \item \emph{action-name}
  \item cas.\emph{action-set.action-name}
}
The third form matches the generated functions for
the CAS actions.}

\item{display}{Should the parameters be printed to the
screen?}
}
\value{
List of action parameters
}
\description{
This function displays the parameters for the specified
CAS action.
}
\examples{
\dontrun{
# specify the action set and action name
.list_action_params(x, action = "simple.summary")

# fetch is in the table action set
.list_action_params(s, action = "fetch")

# specify the generated function name
.list_action_params(s, cas.regression.logistic)
}

}
\keyword{internal}
