% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{.list_action_sets}
\alias{.list_action_sets}
\title{List CAS Action Sets}
\usage{
.list_action_sets(x)
}
\arguments{
\item{x}{An instance of a CAS object that represents
a connection and CAS session.}
}
\value{
List of actionset information
}
\description{
This function lists all action sets that are installed on the
CAS server. The results include a column that is named loaded
to indicate if the action set is already available for use
or must be loaded before it can be used.
}
\examples{
\dontrun{
.list_action_sets(conn)
}

}
\keyword{internal}
