% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.no_null_string}
\alias{.no_null_string}
\title{Utility function for guaranteeing a string result}
\usage{
.no_null_string(str)
}
\arguments{
\item{str}{An arbitrary string}
}
\value{
A string is always returned. If the given value was NULL,
  an empty string is returned; otherwise, the string argument
  given is returned.
}
\description{
Utility function for guaranteeing a string result
}
\keyword{internal}
