% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{.output_file}
\alias{.output_file}
\title{Construct a filename based on the CAS table name}
\usage{
.output_file(x, file, ext)
}
\arguments{
\item{x}{CASTable object}

\item{file}{Input date file name}

\item{ext}{Output file extension}
}
\value{
character
}
\description{
Construct a filename based on the CAS table name
}
\keyword{internal}
