% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_rest.R
\name{.paste_prefix}
\alias{.paste_prefix}
\title{Paste strings together using sep as needed}
\usage{
.paste_prefix(str1, str2, sep = ".")
}
\description{
Paste strings together using sep as needed
}
\keyword{internal}
