% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.r2cas}
\alias{.r2cas}
\title{Convert an R object to a CASValueList}
\usage{
.r2cas(soptions, sw_error, a)
}
\arguments{
\item{soptions}{An soptions string, primarily used for locale information.}

\item{sw_error}{A \code{CASError} object for storing resulting error messages.}

\item{a}{Arbitrary R object to convert.}
}
\description{
Convert an R object to a CASValueList
}
\keyword{internal}
