% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{.run_sas_code}
\alias{.run_sas_code}
\title{Run DATA step action}
\usage{
.run_sas_code(x, code = "")
}
\arguments{
\item{x}{\code{\link{CAS}} connection object.}

\item{code}{DATA step code.}
}
\value{
datastep.runCode action results
}
\description{
Run DATA step action
}
\keyword{internal}
