% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{.set_list_value}
\alias{.set_list_value}
\title{Set a key / value pair in a CASValueList}
\usage{
.set_list_value(sw_values, i, key, value)
}
\arguments{
\item{sw_values}{a \code{CASValueList} object.}

\item{i}{The index where the key / value pair should be inserted.}

\item{key}{String containing the key. This can be an empty string
for indexed lists.}

\item{value}{The value to set.}
}
\value{
The next list index to be populated.
}
\description{
Set a key / value pair in a CASValueList
}
\keyword{internal}
