% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{.summary_stat.CASTable}
\alias{.summary_stat.CASTable}
\title{Function for retrieving a single summary statistic}
\usage{
\method{.summary_stat}{CASTable}(x, stat, na.rm = FALSE, numeric.only = FALSE)
}
\arguments{
\item{x}{CASTable}

\item{stat}{Name of statistic to compute}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Function for retrieving a single summary statistic
}
\keyword{internal}
