% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_rest.R
\name{.trace_actions}
\alias{.trace_actions}
\title{Print tracing information for action call}
\usage{
.trace_actions(action_name, params)
}
\description{
Print tracing information for action call
}
\keyword{internal}
