% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_rest.R
\name{.trace_list}
\alias{.trace_list}
\title{Print tracing information for list parameter}
\usage{
.trace_list(params, prefix = NULL)
}
\description{
Print tracing information for list parameter
}
\keyword{internal}
