% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.translate}
\alias{.translate}
\title{Translate CAS result data frame to list}
\usage{
.translate(df, col = 2L)
}
\description{
This is an internal function to aid in formating results as \R users expect.
}
\keyword{internal}
