% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.unique_table_name}
\alias{.unique_table_name}
\title{Generate a unique table name}
\usage{
.unique_table_name(prefix = "")
}
\arguments{
\item{prefix}{Optional prefix to add to generated name.}
}
\value{
String
}
\description{
Generate a unique table name
}
\keyword{internal}
