% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{drop.table,CASTable-method}
\alias{drop.table,CASTable-method}
\title{Remove a CAS Table}
\usage{
\S4method{drop.table}{CASTable}(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\description{
Drops the in-memory table on the server that is
referenced by the \code{\link{CASTable}} object.
}
\section{Note}{

This function drops the in-memory table but does
not affect the original source file that the
in-memory table was loaded from.
}

\examples{
\dontrun{
drop.table(ct1)
}

}
