% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R
\name{dropTable}
\alias{dropTable}
\title{Remove a CAS Table}
\usage{
dropTable(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\description{
Drops the in-memory table on the server that is
referenced by the \code{\link{CASTable}} object.
}
\section{Note}{

This function drops the in-memory table but does
not affect the original source file that the
in-memory table was loaded from.
}

\examples{
\dontrun{
dropTable(ct1)
}
}
% Copyright SAS Institute
