% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExampleData.R
\name{extracExampleData}
\alias{extracExampleData}
\title{Extract example data}
\usage{
extracExampleData(exampleData, dataName, toDir)
}
\arguments{
\item{exampleData}{is example data, which contains different data, identify
by dataName}

\item{dataName}{name of the data, use one of the following keywords to
extract the data that needed.\cr

"swatTxtInOut" - example TxtInOut folder (and its files) of SWAT \cr
"swatParam" - example parameter file (swatParam.txt) of SWAT \cr
"swatObservedData" - example observed streamflow data (obs_var_1.txt) of SWAT \cr
"swatPlusTxtInOut" - example TxtInOut folder (and its files) of SWAT+ \cr
"swatPlusParam" - example parameter file (cal_parms.cal) of SWAT+ \cr
"swatPlusObservedData" - example observed streamflow data (obs_var_1.txt) of SWAT+ \cr
"all" - extract all data}

\item{toDir}{path to the directory that the example data will be extracted}
}
\value{
the files content will be extracted to the given path
}
\description{
Extract example data
}
\examples{

\donttest{
extracExampleData(exampleData, "swatParam", tempdir())
}

}
\seealso{
\code{\link{exampleData}} for data description
}
