% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHruInfo.R
\name{getHruInfo}
\alias{getHruInfo}
\title{Get information of all HRU (landuse, soil, slope) in all .hru files of SWAT}
\usage{
getHruInfo(TxtInOut)
}
\arguments{
\item{TxtInOut}{Path to TxtInOut folder of SWAT}
}
\value{
a dataframe of HRU information (see example below)
}
\description{
Get information of all HRU (landuse, soil, slope) in all .hru files of SWAT
}
\examples{

\donttest{
# Create a directory and populate with TxtInOut of SWAT
extracExampleData(exampleData,"swatTxtInOut", tempdir())
TxtInOut <- file.path(tempdir(), "swatTxtInOut")

# Now get hru information
getHruInfo(TxtInOut)

}

}
