% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHruInfo.R
\name{getInfoHruHeader}
\alias{getInfoHruHeader}
\title{Extract information from a header of an hru file of SWAT}
\usage{
getInfoHruHeader(hruHeader)
}
\arguments{
\item{hruHeader}{Content of the first line of a hru file}
}
\value{
A list of hru information
}
\description{
Extract information from a header of an hru file of SWAT
}
\examples{

hruHeader <- paste0(' .hru file Watershed HRU:2 Subbasin:1 HRU:2 Luse:FRSE ',
'Soil: Fr333ab4494 Slope: 5-9999 9/21/2021 12:00:00 AM ArcSWAT 2012.10_5.21')

getInfoHruHeader(hruHeader)

}
\keyword{internal}
