% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{getnext}
\alias{getnext}
\title{Return the next response or result from given object}
\usage{
getnext(...)
}
\arguments{
\item{\ldots}{One or more \code{\link{CAS}},
\code{\link{CASEventWatcher}}, or \code{\link{CASResponse}} objects.}
}
\value{
If the arguments given are \code{CAS} or \code{CASEventWatcher}
  objects, the return value is a two-element list with fields \code{response}
  and \code{connection} corresponding to the next retrieved response and
  the connection that it came from. If there are no more respnoses, a NULL
  is returned. If the arguments given, are \code{CASResponse} objects,
  the next value in the response is returned as an R object.
}
\description{
This function is typically used in conjunction with \code{CAS$invoke()}.
It can be used to retrieve responses from multiple CAS action calls
as they become available.
}
\examples{
\dontrun{
  while (TRUE) {
    nextresp <- getnext(.self, datamsghandler = datamsghandler)
    if (is.null(nextresp$response)) break

    while (TRUE) {
      result <- getnext(nextresp$response)
      if (is.null(result)) break

      print(result)
    }
  }
}

}
