% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOutputVar.R
\name{getNumberOutputVar}
\alias{getNumberOutputVar}
\title{Get the number of output variables}
\usage{
getNumberOutputVar(outputExtraction)
}
\arguments{
\item{outputExtraction}{a dataframe instructs which outputs should be
extracted after each run/iteration}
}
\value{
number of output variables (nOutputVar) and boolen variables indicate
whether a specific output is extracted by the user defined extraction function
or not
}
\description{
Get the number of output variables
}
\examples{
# Here is the example of output extraction data frame of SWAt
outputExtractionSWAT <- exampleData$dataOutputExtractionSWAT
outputExtractionSWATPlus <- exampleData$dataOutputExtractionSWATPlus

# Get number of output variables
getNumberOutputVar(outputExtractionSWAT)
getNumberOutputVar(outputExtractionSWATPlus)

}
