% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSWATpar.R
\name{getParameterValue}
\alias{getParameterValue}
\title{Get parameter value}
\usage{
getParameterValue(parameter, parameterValue)
}
\arguments{
\item{parameter}{parameter flag}

\item{parameterValue}{parameter value}
}
\value{
parameter value
}
\description{
Get parameter value
}
\keyword{interal}
