% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamRange.R
\name{getParamRange}
\alias{getParamRange}
\title{Get range (min, max) of selected parameter}
\usage{
getParamRange(paraSelection)
}
\arguments{
\item{paraSelection}{a data frame}
}
\value{
a matrix with 2 columns (min, max) and nrows (= number of parameters)
}
\description{
Get range (min, max) of selected parameter
}
\examples{

# Example of parameter selection dataframe of SWAT and SWAT+
exampleData$paraSelectionSWAT
exampleData$paraSelectionSWATPlus

getParamRange(exampleData$paraSelectionSWAT)
getParamRange(exampleData$paraSelectionSWATPlus)

}
