% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRchNumber.R
\name{getRchNumber}
\alias{getRchNumber}
\title{Get reach/subbasin/reservoir number}
\usage{
getRchNumber(inputText)
}
\arguments{
\item{inputText}{character of the selected reach or subbasins, or resevoir
number}
}
\value{
a list object of output
}
\description{
Get reach/subbasin/reservoir number
}
\examples{

getRchNumber(" 1, 3 * 5,7")

}
