% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{glue}
\alias{glue}
\title{Generalized likelihood uncertainty estimation (GLUE)}
\usage{
glue(perc, x, w)
}
\arguments{
\item{perc}{percentiles}

\item{x}{values (numeric vector). In R-SWAT x is simulated values at a specific
time step (only from behavioral simulations). So to get, for example, the 95PPU
of all time step, need to put this function into a loop.}

\item{w}{weights (numeric vector, sum = 1). This is the objective function values
of behavioral simulations, this must be positive number, current code does not
support negative objective function values when calculating 95PPU using GLUE.
In other works, the behavioral threshold should be positive in this case.}
}
\value{
the value of x at given percentiles
}
\description{
Calculate the percentiles according Generalized likelihood uncertainty
estimation (GLUE). The code was taken from
\url{http://www.uncertain-future.org.uk/wp-content/uploads/2016/06/R-GLUE.zip}
}
\examples{
# Simulated streamflow Qt at day t = t from 100 behavioral simulations
Qt = runif(100)

# NSE of 100 behavioral simulations
NSE = runif(100)
normalize_NSE = NSE/sum(NSE)

# Want to have streamflow at 2.5 and 97.5 percentiles according to GLUE

percentiles = c(0.025, 0.975)

glue(perc=percentiles, x = Qt, w = normalize_NSE)

}
