% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hruSubset.R
\name{hruSubset}
\alias{hruSubset}
\title{Get subset of HRU based on certain criteria}
\usage{
hruSubset(HRUinfo, selectCriteria)
}
\arguments{
\item{HRUinfo}{a dataframe which contains HRU information}

\item{selectCriteria}{a dataframe contains different selection criteria about
landuse (lu), soil (soil), slope (slope), and subbasin (sub) number}
}
\value{
a character vector of hru file names
}
\description{
Get subset of HRU based on certain criteria
}
\examples{

\donttest{
# Create a folder and extract TxtInOut of SWAT there
extracExampleData(exampleData,"swatTxtInOut", tempdir())
TxtInOut <- file.path(tempdir(), "swatTxtInOut")

# Now get hru information
HRUinfo <- getHruInfo(TxtInOut)

# The selection criteria is
selectCriteria <- data.frame(lu = c("AGRL"), soil = c("Fr333ab4494"),
                             sl=c("0-5"), sub =c(1,2))

hruSubset(HRUinfo, selectCriteria)
}

}
\seealso{
\code{\link{getHruInfo}} for how to get HRU information from TxtInOut
}
