% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{is.CASTable}
\alias{is.CASTable}
\title{Test if an object is a CAS Table}
\usage{
is.CASTable(x)
}
\arguments{
\item{x}{An \code{R} object.}
}
\value{
boolean
}
\description{
Test if an object is a CAS Table
}
\examples{
\dontrun{
is.CASTable(ct1) # TRUE
is.CASTable(iris) # FALSE
}

}
