% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{length,CASTable-method}
\alias{length,CASTable-method}
\title{Number of Columns in a CAS Table}
\usage{
\S4method{length}{CASTable}(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
scalar
}
\description{
Returns the number of columns in an in-memory table
that is referenced by the \code{\link{CASTable}} object.
}
\examples{
\dontrun{
length(ct1)
length(ct[1:4])
}

}
\seealso{
\code{ncol,CASTable-method}
}
