% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{listActionSets}
\alias{listActionSets}
\title{List CAS Action Sets}
\usage{
listActionSets(conn)
}
\arguments{
\item{conn}{An instance of a CAS object that represents
a connection and CAS session.}
}
\description{
This function lists all action sets that are installed on the
CAS server. The results include a column that is named loaded
to indicate if the action set is already available for use
or must be loaded before it can be used.
}
\examples{
\dontrun{
listActionSets(conn)
}

}
% Copyright SAS Institute
