% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{loadActionSet}
\alias{loadActionSet}
\title{Load a CAS Action Set}
\usage{
loadActionSet(conn, actionSet = "")
}
\arguments{
\item{conn}{An instance of a CAS object that represents
a connection and CAS session.}

\item{actionSet}{A \code{string} value that specifies
the action set to load.}
}
\description{
This function loads a CAS action set and generate an \R function
for each action.
}
\examples{
\dontrun{
loadActionSet(conn, actionSet = "decisionTree")
loadActionSet(conn, actionSet = "regression")
}

}
