% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSwatParam.R
\name{loadSwatParam}
\alias{loadSwatParam}
\title{Read swatParam.txt file of SWAT or cal_parms.cal of SWAT+}
\usage{
loadSwatParam(swatParamFile)
}
\arguments{
\item{swatParamFile}{swatParam.txt or cal_parms.cal file}
}
\value{
Dataframe of swatParam.txt or cal_parms.cal
}
\description{
Read swatParam.txt file of SWAT or cal_parms.cal of SWAT+
}
\examples{

\donttest{
# Create SWAT parameter file
extracExampleData(exampleData,"swatParam", tempdir())

# Read SWAT prameter file
loadSwatParam(file.path(tempdir(), "swatParam.txt"))

# Create SWAT+ parameter file
extracExampleData(exampleData,"swatPlusParam", tempdir())

# Read SWAT+ prameter file
loadSwatParam(file.path(tempdir(), "cal_parms.cal"))
}

}
