% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{max,CASTable-method}
\alias{max,CASTable-method}
\title{Maximum Values (numeric)}
\usage{
\S4method{max}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Numeric
}
\description{
This method only computes the maximum of numeric values.
}
\examples{
\dontrun{
max(ct1$X1)
max(ct1[1:2])
}
}
\seealso{
\code{cas.max}
}
