% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{mean,CASTable-method}
\alias{mean,CASTable-method}
\title{Mean Value}
\usage{
\S4method{mean}{CASTable}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}

\item{\ldots}{Additional arguments. Currently ignored.}
}
\value{
Numeric
}
\description{
Mean Value
}
\examples{
\dontrun{
mean(ct1$x1)
mean(ct1[2])
}
}
\seealso{
\code{cas.mean}
}
