% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\docType{methods}
\name{median,CASTable-method}
\alias{median,CASTable-method}
\title{Median Values}
\usage{
\S4method{median}{CASTable}(x)
}
\arguments{
\item{x}{CASTable.}
}
\value{
scalar
}
\description{
Median Values
}
\examples{
\dontrun{
median(ct1$x1)
median(ct1[1:4])
}
}
\seealso{
\code{cas.median}
}
% Copyright SAS Institute
