% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{names<-,CASTable-method}
\alias{names<-,CASTable-method}
\title{Set names of a CAS Table}
\usage{
\S4method{names}{CASTable}(x) <- value
}
\arguments{
\item{x}{A CASTable object.}
}
\description{
Set names of a CAS Table
}
