% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R
\docType{methods}
\name{ncol,CASTable-method}
\alias{ncol,CASTable-method}
\title{Number of Columns in a CAS Table}
\usage{
\S4method{ncol}{CASTable}(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
scalar
}
\description{
Returns the number of columns in an in-memory table
that is referenced by the \code{\link{CASTable}} object.
}
\examples{
\dontrun{
ncol(ct1)
ncol(ct[1:4])
}
}
\seealso{
\code{length,CASTable-method}
}
% Copyright SAS Institute
