% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R
\name{nrow,CASTable-method}
\alias{nrow,CASTable-method}
\title{Number of Rows in a CAS Table}
\usage{
\S4method{nrow}{CASTable}(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
scalar
}
\description{
Returns the number of rows in an in-memory table
that is referenced by the \code{\link{CASTable}} object.
}
\examples{
\dontrun{
nrow(ct1)
nrow(ct[1:4])
}
}
% Copyright SAS Institute
