% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calObjFunction.R
\name{observedToList}
\alias{observedToList}
\title{Subset of observed data according to flag}
\usage{
observedToList(observedData, flag)
}
\arguments{
\item{observedData}{list of observed data, each observed data is a data frame}

\item{flag}{character ("C" for calibration and "V" for validation)}
}
\value{
list of calibrated or validated data
}
\description{
Subset of observed data according to flag
}
\examples{
\donttest{
observedToList(observedData, flag)
}

}
\keyword{internal}
