% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calObjFunction.R
\name{perCriteria}
\alias{perCriteria}
\title{Calculate NSE, KGE, RMSE, R2, and aBIAS}
\usage{
perCriteria(obs, sim)
}
\arguments{
\item{obs}{vector of observed data (NA values are allowed)}

\item{sim}{vector of simulated data (NA values are allowed)}
}
\value{
dataframe with 5 columns (NSE, KGE, RMSE, R2, and aBIAS) and 1 row
}
\description{
Calculate NSE, KGE, RMSE, R2, and aBIAS
}
\examples{

perCriteria(obs=runif(10), sim=runif(10))

}
