% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{POSIXct.as.CASdt}
\alias{POSIXct.as.CASdt}
\title{Convert a POSIXct to a CAS timestamp}
\usage{
POSIXct.as.CASdt(rpct)
}
\arguments{
\item{rpct}{POSIXct object}
}
\value{
Numeric representing a CAS timestamp
}
\description{
Convert a POSIXct to a CAS timestamp
}
