% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{POSIXct.as.SASdt}
\alias{POSIXct.as.SASdt}
\title{Convert a POSIXct to a SAS timestamp}
\usage{
POSIXct.as.SASdt(rpct)
}
\arguments{
\item{rpct}{POSIXct object}
}
\value{
Numeric representing a SAS timestamp
}
\description{
Convert a POSIXct to a SAS timestamp
}
