% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{POSIXlt.as.CASdt}
\alias{POSIXlt.as.CASdt}
\title{Convert a POSIXlt to a CAS timestamp}
\usage{
POSIXlt.as.CASdt(rplt)
}
\arguments{
\item{rplt}{POSIXlt object}
}
\value{
Numeric representing a CAS timestamp
}
\description{
Convert a POSIXlt to a CAS timestamp
}
