% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{POSIXlt.as.SASdt}
\alias{POSIXlt.as.SASdt}
\title{Convert a POSIXlt to a SAS timestamp}
\usage{
POSIXlt.as.SASdt(rplt)
}
\arguments{
\item{rplt}{POSIXlt object}
}
\value{
Numeric representing a SAS timestamp
}
\description{
Convert a POSIXlt to a SAS timestamp
}
