% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prFactor.R
\name{prFactor}
\alias{prFactor}
\title{Calculate p-factor and r-factor}
\usage{
prFactor(obs, low, up)
}
\arguments{
\item{obs}{observed data (vector)}

\item{low}{lower values of the 95PPU (at 2.5\%)}

\item{up}{upper values of the 95PPU (at 97.5\%)}
}
\value{
vector of two values c(p-factor, r-factor)
}
\description{
Calculate p-factor and r-factor
}
\examples{

\donttest{
# Create aritifical simulated streamflow (95PPU) and observed streamflow
Q_sim_95ppu_low <-  2 + sin(seq(0,10,0.1))
Q_sim_95ppu_high <- 2.2 + sin(seq(0,10,0.1))
Q_obs <- (Q_sim_95ppu_low + Q_sim_95ppu_high)/2 + runif(101)*0.2

# Plot to see data
plot(Q_sim_95ppu_low, col = "black", type = "l", xlab = "Time (days)",
     ylab = "Streamflow (m3/s)")
lines(Q_sim_95ppu_high, col = "red")
points(Q_obs, col = "blue")

# Calculate p and r factor
prFactor(Q_obs, Q_sim_95ppu_low, Q_sim_95ppu_high )
}

}
