% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{quantile,CASTable-method}
\alias{quantile,CASTable-method}
\title{Quantile and Percentile Values}
\usage{
\S4method{quantile}{CASTable}(
  x,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = TRUE,
  type = 7,
  ...
)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{probs}{Numeric vector of probabilities with values in [0,1].}

\item{na.rm}{Should missing values be omitted from the calculations?}

\item{\ldots}{Additional arguments. Currently ignored.}
}
\value{
Numeric
}
\description{
Quantile and Percentile Values
}
\examples{
\dontrun{
quantile(ct1[2])
quantile(ct1$x1, probs = c(10, 90))
}

}
\seealso{
\code{cas.quantile}
}
