% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{rbind,CASTable-method}
\alias{rbind,CASTable-method}
\title{Combine CAS Tables by Rows}
\usage{
\S4method{rbind}{CASTable}(..., deparse.level = 1)
}
\arguments{
\item{\dots}{Arguments that are passed to method arguments.}

\item{deparse.level}{See the help for base::rbind.}
}
\value{
\code{\link{CASTable}}
}
\description{
This is the implementation of rbind for in-memory tables.
}
\examples{
\dontrun{
rbind(ct1, ct2)
rbind(ct1, ct2, ct3)
}

}
