% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bygroups.R
\name{rbind.bygroups,list-method}
\alias{rbind.bygroups,list-method}
\title{Combine data.frames belonging to the same By Group set into one data.frame}
\usage{
\S4method{rbind.bygroups}{list}(x)
}
\arguments{
\item{x}{List containing CAS action results.}
}
\value{
List of By Group sets with a single data.frame in each field.
}
\description{
Combine data.frames belonging to the same By Group set into one data.frame
}
