% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{rbind.bygroups}
\alias{rbind.bygroups}
\title{Combine data.frames belonging to the same By Group set into one data.frame}
\usage{
rbind.bygroups(res)
}
\arguments{
\item{res}{List containing CAS action results.}
}
\value{
List of By Group sets with a single data.frame in each field.
}
\description{
Combine data.frames belonging to the same By Group set into one data.frame
}
