% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{rbind.casTable}
\alias{rbind.casTable}
\title{Combine CAS Tables by Columns}
\usage{
rbind.casTable(..., deparse.level = 1)
}
\arguments{
\item{\dots}{Arguments that are passed to method arguments.}

\item{deparse.level}{See the help for base::rbind.}
}
\description{
Combine CAS Tables by Columns
}
