% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\docType{methods}
\name{rbind2.casTable}
\alias{rbind2.casTable}
\alias{rbind2,CASTable,ANY-method}
\title{Combine CAS Tables by Rows}
\usage{
rbind2.casTable(x, y, ...)

\S4method{rbind2}{CASTable,ANY}(x, y, ...)
}
\arguments{
\item{\dots}{Arguments that are passed to method arguments.}

\item{deparse.level}{See the help for base::rbind.}
}
\value{
\code{\link{CASTable}}
}
\description{
This is the implementation of \code{rbind} for in-memory tables.

Combine CAS Tables by Columns
}
\examples{
\dontrun{
rbind(ct1,ct2)
rbind(ct1,ct2, ct3)
}
}
% Copyright SAS Institute
