% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChannelFile.R
\name{readChannelFile}
\alias{readChannelFile}
\title{Read channel_sd... files and lsunit_wb... files of SWAT+}
\usage{
readChannelFile(
  workingDirectory,
  coreNumber,
  fileName,
  fromToDate,
  colNumber,
  fileCioInfo,
  rchNumber,
  output
)
}
\arguments{
\item{workingDirectory}{path to the working directory where all files created
by SWAT or SWAT+ will be saved}

\item{coreNumber}{which core number you are reading from}

\item{fileName}{which file name (output.rch, output.sub or output.hru)}

\item{fromToDate}{a date vector with length of 2, indicating from which time
period the function should extract outputs}

\item{colNumber}{which column number of the file}

\item{fileCioInfo}{a dataframe containing simulation time, warum up time}

\item{rchNumber}{which reach, or subbasin, or hru number}

\item{output}{a list object containing another outputs and outputs from calling
this function}
}
\value{
no return value, results will be saved in the workingDirectory
}
\description{
Read channel outputs (any file starts with channel_sd) and landscape unit
file (any file starts with lsuni_wd) of SWAT+
}
\examples{

# Please see in the

}
