% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFileContent.R
\name{readFileContent}
\alias{readFileContent}
\title{Read content of multiple files}
\usage{
readFileContent(fileDirectory, files)
}
\arguments{
\item{fileDirectory}{a directory where these files are located}

\item{files}{a character vector of file names (only file names, without
full path) that will be read}
}
\value{
a list store the content of each file
}
\description{
Read content of multiple files
}
\examples{

\donttest{
readFileContent(fileDirectory, files)
}
}
\keyword{internal}
