% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSaveOutput.R
\name{readOutputRchFile}
\alias{readOutputRchFile}
\title{Read output from output.rch, output.sub, or output.hru files of SWAT}
\usage{
readOutputRchFile(
  workingDirectory,
  coreNumber,
  fileName,
  fromToDate,
  colNumber,
  fileCioInfo,
  rchNumber,
  output
)
}
\arguments{
\item{workingDirectory}{path to the working directory where all files created
by SWAT or SWAT+ will be saved}

\item{coreNumber}{which core number you are reading from}

\item{fileName}{which file name (output.rch, output.sub or output.hru)}

\item{fromToDate}{a date vector with length of 2, indicating from which time
period the function should extract outputs}

\item{colNumber}{which column number of the file}

\item{fileCioInfo}{a dataframe containing simulation time, warum up time}

\item{rchNumber}{which reach, or subbasin, or hru number}

\item{output}{a list object that results should be attached to}
}
\value{
a list object of output
}
\description{
Read output from output.rch, output.sub, or output.hru files of SWAT
}
