% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWatoutFile.R
\name{readWatoutFile}
\alias{readWatoutFile}
\title{Read watout.dat file from SWAT}
\usage{
readWatoutFile(
  workingDirectory,
  coreNumber,
  fileName,
  fromToDate,
  colNumber,
  fileCioInfo,
  output
)
}
\arguments{
\item{workingDirectory}{path to the working directory where all files created
by SWAT or SWAT+ will be saved}

\item{coreNumber}{which core number you are reading from}

\item{fileName}{which file name (output.rch, output.sub or output.hru)}

\item{fromToDate}{a date vector with length of 2, indicating from which time
period the function should extract outputs}

\item{colNumber}{which column number of the file}

\item{fileCioInfo}{a dataframe containing simulation time, warum up time}

\item{output}{a list object that results should be attached to}
}
\value{
a list object of output
}
\description{
Read watout.dat file of SWAT. This file is always at daily timestep and
include the warmup period
}
