% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_rest.R
\docType{class}
\name{REST_CASConnectionEventWatcher-class}
\alias{REST_CASConnectionEventWatcher-class}
\alias{REST_CASConnectionEventWatcher}
\title{CAS connection event watcher class}
\description{
CAS connection event watcher class
}

\keyword{internal}
