% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R
\name{rownames,CASTable-method}
\alias{rownames,CASTable-method}
\title{Row Names of a CAS Table}
\usage{
\S4method{rownames}{CASTable}(x)
}
\arguments{
\item{x}{A \code{\link{CASTable}} object.}
}
\value{
list of strings
}
\description{
Row Names of a CAS Table
}
\examples{
\dontrun{
rownames(ct1)
rownames(ct[2:4])
}
}
% Copyright SAS Institute
