% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{runAction}
\alias{runAction}
\title{Run a CAS Action by Name}
\usage{
runAction(CASorCASTab = "", actn, ...)
}
\arguments{
\item{CASorCASTab}{An instance of a CAS object that represents
a connection and CAS session, or an instance of a CASTable.}

\item{actn}{A \code{character} string that specifies
the action set and action name to run.}

\item{\dots}{Parameters that are passed to the CAS action.}
}
\description{
This function enables you to specify a CAS action
name and run it. This is an alternative to running
an action from the generated function names. An
example of a generated function name is cas.table.tableInfo.
}
\examples{
\dontrun{
# display the active caslib for the session
runAction(conn, "sessionProp.getSessOpt", name="caslib")

# display simple summary statistics for an uploaded data frame
mtcarsct <- as.casTable(conn, mtcars)
runAction(mtcarsct, "simple.summary")

# the preceding runAction function is functionally
# equivalent to the following generated function
cas.simple.summary(mtcarsct)
}

}
% Copyright SAS Institute
