% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASd.as.CASd}
\alias{SASd.as.CASd}
\title{Convert a SAS date to a CAS date}
\usage{
SASd.as.CASd(sdt)
}
\arguments{
\item{sdt}{Numeric representing a SAS date}
}
\value{
Numeric representing a CAS date
}
\description{
Convert a SAS date to a CAS date
}
