% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASd.as.Date}
\alias{SASd.as.Date}
\title{Convert a SAS date to a Date}
\usage{
SASd.as.Date(sdt)
}
\arguments{
\item{sdt}{Numeric representing a SAS date}
}
\value{
Date object
}
\description{
Convert a SAS date to a Date
}
