% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASdt.as.CASdt}
\alias{SASdt.as.CASdt}
\title{Convert a SAS timestamp to a CAS timestamp}
\usage{
SASdt.as.CASdt(sts)
}
\arguments{
\item{sts}{Numeric representing a SAS timestamp}
}
\value{
Numeric representing a CAS timestamp
}
\description{
Convert a SAS timestamp to a CAS timestamp
}
