% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASdt.as.POSIXct}
\alias{SASdt.as.POSIXct}
\title{Convert a SAS timestamp to a POSIXct}
\usage{
SASdt.as.POSIXct(sts)
}
\arguments{
\item{sts}{Numeric representing a SAS timestamp}
}
\value{
POSIXct object
}
\description{
Convert a SAS timestamp to a POSIXct
}
