% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASdt.as.POSIXlt}
\alias{SASdt.as.POSIXlt}
\title{Convert a SAS timestamp to a POSIXlt}
\usage{
SASdt.as.POSIXlt(sts)
}
\arguments{
\item{sts}{Numeric representing a SAS timestamp}
}
\value{
POSIXlt object
}
\description{
Convert a SAS timestamp to a POSIXlt
}
