% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASt.as.CASt}
\alias{SASt.as.CASt}
\title{Convert a SAS time to a CAS time}
\usage{
SASt.as.CASt(stm)
}
\arguments{
\item{stm}{Numeric representing a SAS time}
}
\value{
Numeric representing a CAS time
}
\description{
Convert a SAS time to a CAS time
}
