% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASt.as.POSIXct}
\alias{SASt.as.POSIXct}
\title{Convert a SAS time to a POSIXct}
\usage{
SASt.as.POSIXct(stm)
}
\arguments{
\item{stm}{Numeric representing a SAS time}
}
\value{
POSIXct object
}
\description{
Convert a SAS time to a POSIXct
}
