% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{SASt.as.POSIXlt}
\alias{SASt.as.POSIXlt}
\title{Convert a SAS timem to a POSIXlt}
\usage{
SASt.as.POSIXlt(stm)
}
\arguments{
\item{stm}{Numeric representing a SAS time}
}
\value{
POSIXlt object
}
\description{
Convert a SAS timem to a POSIXlt
}
