% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{saveOutput}
\alias{saveOutput}
\title{Save output of SWAT/SWAT+ after each model run}
\usage{
saveOutput(
  workingDirectory,
  coreNumber,
  fileName,
  fileType,
  fromToDate,
  colNumber,
  rchNumber,
  fileCioInfo,
  simulationNumber,
  firstRun
)
}
\arguments{
\item{workingDirectory}{path to the working directory where all files created
by SWAT or SWAT+ will be saved}

\item{coreNumber}{which core number you are reading from}

\item{fileName}{which file name (output.rch, output.sub or output.hru)}

\item{fileType}{character, type of the file that are being read to extract
output, possible values are:
"watout.dat" \cr
"output.rch" \cr
"output.sub" \cr
"output.hru" \cr
"channel_sd_day.txt" \cr
"channel_sd_mon.txt" \cr
"channel_sd_yr.txt" \cr
"channel_sdmorph_day.txt"  \cr
"channel_sdmorph_mon.txt" \cr
"channel_sdmorph_yr.txt" \cr
"lsunit_wb_day.txt"  \cr
"lsunit_wb_mon.txt" \cr
"lsunit_wb_yr.txt"  \cr
"basin_wb_day.txt" \cr
"basin_wb_mon.txt" \cr
"basin_wb_yr.txt" \cr}

\item{fromToDate}{a date vector with length of 2, indicating from which time
period the function should extract outputs}

\item{colNumber}{which column number of the file}

\item{rchNumber}{which reach, or subbasin, or hru number}

\item{fileCioInfo}{a dataframe containing simulation time, warum up time}

\item{simulationNumber}{integer, the simulation number (not the number of
simulations)}

\item{firstRun}{logical (set to TRUE if this is the first iteration/run),
in the first iteration, all files need to be copied to the new TxtInOut folder
for SWAT/SWAT+ run, in the subsequent iterations, only files that content
parameters (that need to be changed) are updated.}
}
\value{
no return value, results will be saved in the workingDirectory
}
\description{
Save output of SWAT/SWAT+ after each model run
}
