% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{show,CASTable-method}
\alias{show,CASTable-method}
\title{Print a sample of the CASTable to the screen}
\usage{
\S4method{show}{CASTable}(object)
}
\arguments{
\item{object}{\code{\link{CASTable}} object.}
}
\description{
Print a sample of the CASTable to the screen
}
