% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRSWAT.R
\name{showRSWAT}
\alias{showRSWAT}
\title{Display the Graphical User Interface (GUI) of R-SWAT}
\usage{
showRSWAT()
}
\value{
GUI of R-SWAT
}
\description{
Calling this functions will show the GUI of R-SWAT
}
\examples{

\donttest{
showRSWAT()
}

}
