% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calObjFunction.R
\name{simToList}
\alias{simToList}
\title{Subset of simulated data according to flag}
\usage{
simToList(simData, k, observedData, flag)
}
\arguments{
\item{k}{column number of simulated data}

\item{observedData}{list of observed data, each observed data is a data frame}

\item{flag}{character ("C" for calibration and "V" for validation)}
}
\value{
list of simulated data  for the calibration or validation period
}
\description{
Subset of simulated data according to flag
}
\examples{
\donttest{
simToList(simData, k, observedData, flag)
}


}
\keyword{internal}
