% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortObservedDataFile.R
\name{sortObservedDataFile}
\alias{sortObservedDataFile}
\title{Sort observed data files by it number before the file extention}
\usage{
sortObservedDataFile(observedDataFile)
}
\arguments{
\item{observedDataFile}{a character vector, containing the NAME different
observed data files, all observed data files must have the name starting with
'obs_var_', followed by an integer number, indicating the variable number, and
the '.txt' extension. Please see example below.}
}
\value{
a character vector, containing the sorted names of observed data files
}
\description{
Sort observed data files by it number before the file extention
}
\examples{

observedDataFile <- c("obs_var_2.txt", "obs_var_3.txt", "obs_var_1.txt")

sortObservedDataFile(observedDataFile)

}
