% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subParameterSet.R
\name{splitParameterValue}
\alias{splitParameterValue}
\title{Split parametersets (simulations) among threads}
\usage{
splitParameterValue(ncores, parameterValue)
}
\arguments{
\item{ncores}{number of threads for parallel execution}

\item{parameterValue}{a dataframe contains all parametersets for running SWAT
or SWAT+}
}
\value{
A list contain different subset of the parametersets
}
\description{
When run SWAT/SWAT+ parallel, this function will split the parameterset to each
core, for example if there is the number of parametersets is nparams and the
number of cores/threads is ncores. Then each cores will performe nparams/ncores
simulations, if nparams is not devisable to ncores, then the last cores will
take all remaining parametersets
}
\examples{

ncores <- 3
parameterValue <- matrix(runif(20), ncol=2)
splitParameterValue(ncores, parameterValue)

}
