% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{[,CASTable-method}
\alias{[,CASTable-method}
\title{Return a CASTable with filtering applied}
\usage{
\S4method{[}{CASTable}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{i}{Row accessor specification.}

\item{j}{Column accessor specification.}

\item{drop}{Logical.}

\item{\ldots}{Additional arguments. Currently, not used.}
}
\value{
New CASTable object with filtering options applied.
}
\description{
Return a CASTable with filtering applied
}
