% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\name{[[,CASTable-method}
\alias{[[,CASTable-method}
\title{Return a CASTable with filtering applied}
\usage{
\S4method{[[}{CASTable}(x, i)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{i}{Numeric or character item accessor.}
}
\value{
\code{\link{CASTable}} object.
}
\description{
Return a CASTable with filtering applied
}
