% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R, R/table_functions.R
\name{[<-,CASTable-method}
\alias{[<-,CASTable-method}
\alias{[,CASTable-method}
\alias{subset,CASTable-method}
\title{Subset assignment}
\usage{
\S4method{[}{CASTable}(x, i, j, ...) <- value

\S4method{subset}{CASTable}(x, subset, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{i}{Row accessor object.}

\item{j}{Column accessor object}

\item{value}{\code{\link{CASTable}} object or character vector.}

\item{subset}{Logical expression indicating elements or rows to keep.}

\item{select}{Expression indicating columns to select from CAS table.}

\item{drop}{Passed to the \code{[} indexing operator.}

\item{\ldots}{Additional arguments.}
}
\value{
\code{\link{CASTable}} object.
}
\description{
Return a subset of rows and columns from a \code{\link{CASTable}} that meet
subsetting criteria.
}
\examples{
\dontrun{
subset(ct, subset = ct[4] > 15, select = c("n1", "n4", "s"), drop = FALSE)
subset(ct, subset = ct$n4 > 15, select = c(1, 4, 5), drop = FALSE)
}

}
