% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{subset.casTable}
\alias{subset.casTable}
\title{Return a Subset of Rows and Columns from a CAS Table}
\usage{
subset.casTable(x, subset, select = NULL, drop = FALSE, ...)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
A CASTable object with the rows and columns that meet the subset criteria.
}
\description{
Return a subset of rows and columns from a \code{\link{CASTable}} that meet
subsetting criteria.
}
\examples{
\dontrun{
subset(ct, subset = ct[4] > 15, select = c("n1", "n4", "s"), drop = FALSE)
subset(ct, subset = ct$n4 > 15, select = c(1, 4, 5), drop = FALSE)
}
}
% Copyright SAS Institute
