% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{sum,CASTable-method}
\alias{sum,CASTable-method}
\title{Sum of All Values}
\usage{
\S4method{sum}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}

\item{\ldots}{Additional arguments. Currently ignored.}
}
\value{
Numeric
}
\description{
Sum of All Values
}
\examples{
\dontrun{
sum(ct1)
sum(ct1$x1)
}

}
\seealso{
\code{cas.sum}
}
