% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{summary,CASTable-method}
\alias{summary,CASTable-method}
\title{Summary Statistics}
\usage{
\S4method{summary}{CASTable}(object, maxsum = 7, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{\code{\link{CASTable}} object.}

\item{maxsum}{Integer indicating how many levels should be shown
for factors.}

\item{digits}{Integer used for number formatting.}

\item{\ldots}{Additional arguments. Currently ignored.}
}
\value{
table
}
\description{
Returns simple descriptive statistics.
}
\examples{
\dontrun{
summary(ct1)
summary(ct1[1:4])
}
}
% Copyright SAS Institute
