% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{swat.as.integer64}
\alias{swat.as.integer64}
\title{Convert a value to an int64}
\usage{
swat.as.integer64(value)
}
\description{
This is an internal function to convert an arbitrary value to
an int64 value. The current implementation simply returns a
numeric since int64 values are not well supported in R at this time.
}
\keyword{internal}
