% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\docType{package}
\name{swat}
\alias{swat}
\alias{swat-package}
\title{SWAT: SAS Wrapper for Analytics Transfer (SWAT)}
\description{
This package enables you to connect from \R to a SAS Cloud Analytic Services
host, run actions on in-memory tables, and work with the results of the 
actions. 
\itemize{
  \item The \code{\link{CAS}} class provides an interface to your
        connection to the CAS server and CAS session.
  \item The \code{\link{CASTable}} class provides an interface to the
        in-memory tables.
  \item The \code{\link{casDataFrame}} class provides an interface
        to the results for most actions.
}
}
\details{
Depending on how you install the package, you might be able to use binary
communication with CAS. This is more efficient for bandwidth, but requires
that your \R installation have access to a precompiled library (rswat.so or
rswat.dll).  An alternative is to communicate with the server using the REST
interface of the server over HTTP. See the connection examples that follow.

The responses and results of the actions are returned as \R objects.
}
\section{Connect and start a session}{

\preformatted{
  s  <- CAS('cloud.example.com', 5570)  # binary communication

  s2 <- CAS('cloud.example.com', 8777, protocol='https')
}
}

\section{Run a simple action}{

\preformatted{
   results <- runAction(s, "builtins.serverStatus")
   results$server
   nodes actions
 1     1      15
}

You can also run an action using the generated \R function:
\preformatted{
   results <- cas.builtins.serverStatus(s)
   results$server
   nodes actions
 1     1      15
}
}

\section{Upload a data.frame to a CASTable}{

\preformatted{
  irisct <- as.casTable(s, iris)
}
}

\section{Load a CAS actionSet}{

\preformatted{
  runAction(s, "builtins.loadActionSet", actionSet="regression")
}
}

\section{Useful links}{

\itemize{
  \item{\url{http://developer.sas.com/guides/r.html}}
  \item{\url{https://github.com/sassoftware/r-swat}}
  \item{Enter issues at \url{https://github.com/sassoftware/r-swat/issues}}
}
}

\section{Action documentation}{

See the \href{http://documentation.sas.com/?softwareId=pgm
&softwareVersion=production.a&softwareContextId=allProdsActions&showBanner=develop}{
Actions and Action Sets by Name and Product}
}

\author{
\strong{Maintainer}: Jared Dean \email{Jared.Dean@sas.com}

Authors:
\itemize{
  \item Tom Weber \email{Tom.Weber@sas.com}
  \item Kevin Smith \email{Kevin.Smith@sas.com}
}

}
% Copyright SAS Institute
