% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{tail,CASTable-method}
\alias{tail,CASTable-method}
\title{Return the Last Part of a CAS Table}
\usage{
\S4method{tail}{CASTable}(x, n = 6L)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
A \code{\link{casDataFrame}} object with the last n rows.
}
\description{
Returns the last part of an in-memory table that is referenced
by a \code{\link{CASTable}} object.
}
\section{Note}{


The \code{tail} function is not deterministic between reloads of data
or if nodes or added or removed from a distributed server.
}

\examples{
\dontrun{
tail(ct1)
tail(ct[1:4], 10)
}
}
% Copyright SAS Institute
