% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASdf.R
\name{to.r.data.frame}
\alias{to.r.data.frame}
\title{Convert a CAS Table to a R Data Frame (Download)}
\usage{
to.r.data.frame(ct, obs = 32768)
}
\arguments{
\item{ct}{The CASTable object to download.}

\item{obs}{Number of rows to download, by default 32768}
}
\value{
Returns a casDataFrame object that contains
        a copy of the in-memory data.
}
\description{
Downloads the in-memory table that is referenced by
the CASTable object and stores it as a casDataFrame
in R. This function is used to download datasets from CAS.
}
\examples{
\dontrun{
r_df = to.r.data.frame(CASTable)
}

}
% Copyright SAS Institute
