% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{unique,CASTable-method}
\alias{unique,CASTable-method}
\title{Extract Unique Values from a CAS Table}
\usage{
\S4method{unique}{CASTable}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{incomparables}{A vector of values that cannot be compared.
See the help for base::unique.}

\item{\dots}{Arguments that are passed to method arguments.}
}
\value{
A \code{\link{casDataFrame}} object.
}
\description{
Extracts distinct values from columns in a \code{\link{CASTable}}.
}
\examples{
\dontrun{
unique(ct[4:5])
unique(ct$s)
unique(ct[4])
}

}
