% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCalibrationFile.R
\name{updateCalibrationFile}
\alias{updateCalibrationFile}
\title{Update calibration.cal file for SWAT+}
\usage{
updateCalibrationFile(paraSelection, parameterValue, currentDirectory)
}
\arguments{
\item{paraSelection}{a data frame}

\item{parameterValue}{a dataframe contains all parametersets for running SWAT
or SWAT+}

\item{currentDirectory}{path which the new calibration.cal will be saved}
}
\value{
No return values, the new calibration.cal file will be save to
currentDirectory
}
\description{
Update calibration.cal file for SWAT+
}
