% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSWATpar.R
\name{updateMultiFile}
\alias{updateMultiFile}
\title{Update multiple TxtInOut files with new parameter values}
\usage{
updateMultiFile(toDir, caliParam, parameterValue, paraSelection)
}
\arguments{
\item{toDir}{path to the directory where the updated files will be saved}

\item{caliParam}{dataframe of parameters, number of row is number of samples,
the first column is the sample ID and the number of subsequence columns are
the number of parameters.}

\item{parameterValue}{a dataframe contains all parametersets for running SWAT
or SWAT+}

\item{paraSelection}{a data frame}
}
\value{
no return, results will be saved in todir directory
}
\description{
Rewrite the TxtInOut with new parameter values
}
\examples{

\donttest{
updateMultiFile(toDir, caliParam, parameterValue, paraSelection)
}

}
