% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSWATpar.R
\name{updateSingleFile}
\alias{updateSingleFile}
\title{Update multiple TxtInOut files with new parameter values}
\usage{
updateSingleFile(
  toDir,
  file,
  fileContent,
  atLine,
  atPosition,
  changeMethod,
  applyValue,
  numberFormat,
  absoluteMin,
  absoluteMax
)
}
\arguments{
\item{toDir}{path to the directory where the updated files will be saved}

\item{file}{name of the file which contains the updated parameter}

\item{fileContent}{content of the file which contains the updated parameter}

\item{atLine}{line number of the updated parameter}

\item{atPosition}{column location of the updated parameter}

\item{changeMethod}{method of change}

\item{applyValue}{the value which the original parameter should be modify}

\item{numberFormat}{format of the updated parameter}

\item{absoluteMin}{minimum value which the updated parameter should be greater}

\item{absoluteMax}{maximum value which the updated parameter should be exceed}
}
\value{
no return, results will be saved in todir directory
}
\description{
Rewrite the TxtInOut with new parameter values
}
\examples{

\donttest{
updateSingleFile(toDir,
                 file,
                 fileContent,
                 atLine,
                 atPosition,
                 changeMethod,
                 applyValue,
                 numberFormat,
                 absoluteMin,
                 absoluteMax)
}

}
