% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userObjFunction.R
\name{userObjFunction}
\alias{userObjFunction}
\title{User-defined objective function - DO NOT USE AS IT IS NOW}
\usage{
userObjFunction(obs, sim)
}
\arguments{
\item{obs}{a list of observed data, for example, time series of the first
observed variable is obs[[1]], of the second observed variable is obs[[2]],
and so on...}

\item{sim}{a list of simulated data, sim[[1]] corresponds to obs[[1]], and so
on...}
}
\value{
a data frame contains a single objective function value
}
\description{
This is a user-defined objective function, do not use as it is now, please
modify the source code of this function if you use according to your need and
then install RSWAT app again from your local folder to update this function.
DO NOT CHANGE the name of this function (mus be userObjFunction). ONLY CHANGE
the code at the place where indicate "START TO CHANGE FROM HERE" and STOP
at "END OF CHANGE"
}
\examples{

\donttest{
userObjFunction(obs, sim)
}

}
