% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyOutputLoc.R
\name{yearlyOutputLoc}
\alias{yearlyOutputLoc}
\title{Find location of yearly output summary in output.rch, output.sub, output.hru}
\usage{
yearlyOutputLoc(timeSeries, nRch)
}
\arguments{
\item{timeSeries}{vector of data in format (YYYY-MM-DD)}

\item{nRch}{total number of reaches/subbasins/hrus}
}
\value{
location of yearly summary outputs
}
\description{
In files output.rch/.sub/.hru of SWAT, the outputs are summarized every year
and at the end of the simulation, therefore, when reading time series from
these files, we need to skip these rows. This function will find these rows
}
\examples{

timeSeries = seq(as.Date("2000-01-01", "\%Y-\%m-\%d"),
                 as.Date("2005-01-01", "\%Y-\%m-\%d"),
                 by='days')

nRch = 3
yearlyOutputLoc(timeSeries, nRch)

}
