% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\docType{class}
\name{CAS-class}
\alias{CAS-class}
\alias{CAS}
\title{CAS Object Class}
\arguments{
\item{hostname}{A \code{character} string that specifies the 
host name of the CAS controller.}

\item{port}{A \code{numeric} value that specifies the network
port number that the CAS controller listens on.}

\item{protocol}{A \code{string} that specifies one of the following:
\describe{
  \item{cas}{use binary communication. This is the default.}
  \item{http}{use HTTP communication with the REST interface
    on the CAS controller.}
  \item{https}{use HTTPS communication with the REST interface
    on the CAS controller. This protocol must be specified
    explicity.}
  \item{auto}{automatically detect between the binary and HTTP.} 
}}

\item{username}{A \code{character} string that identifies the
user ID to authenticate as.}

\item{password}{A \code{character} string that specifies your
password.}

\item{session}{A \code{character} string that identifies the
32 character UUID of an existing session, if you want to
connect to an existing session. This is rare.}

\item{locale}{A \code{character} string that specifies the
locale to use for the CAS session. By default, the locale
is set to the locale of the server.}

\item{authinfo}{A \code{character} string that specifies
an alternative path to a .authinfo file that is used
for authentication.  By default, ~/.authinfo is used on
Linux and \%HOMEDRIVE\% \\\%HOMEPATH\%\\_authinfo is used on
Windows.}

\item{path}{Base path of the connection URL}

\item{authcode}{Authorization code from SASLogon used to retrieve
an OAuth token.}
}
\value{
A CAS object.
}
\description{
An instance of this class represents a connection and session
between the client (R) and the server (SAS Cloud Analytic Services).
}

\examples{
\dontrun{
# Use binary communication and credentals from the default authinfo location.
s <- CAS('cloud.example.com', 5570) 

# Use HTTPS and credentials from the default authinfo location.
s <- CAS('cloud.example.com', 8777, protocol='https')

# Use binary or HTTP communication and credentials from an alternative authinfo.
s <- CAS('cloud.example.com', 5570, protocol='auto', authinfo=~/alt.txt)

# Use binary or HTTP communcation and specify credentials.
s <- CAS('cloud.example.com', 8777, protocol='auto', username="sasdemo"
  password="!s3cret")
}
}
% Copyright SAS Institute
