% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.close}
\alias{cas.close}
\title{Close a CAS connection while leaving the session alive}
\usage{
cas.close(conn)
}
\arguments{
\item{CAS}{The CAS connection object}
}
\description{
Close a CAS connection while leaving the session alive
}
