% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swat.R
\name{cas.upload.file}
\alias{cas.upload.file}
\title{Upload a data file to a CAS table}
\usage{
cas.upload.file(conn, ...)
}
\arguments{
\item{\dots}{Optional parameters that are passed to the table.loadtable action.}

\item{CAS}{The CAS connection object}

\item{character}{The filename to upload}
}
\value{
\code{\link{CASTable}}
}
\description{
Upload a data file to a CAS table
}
