% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{cas.write.csv2}
\alias{cas.write.csv2}
\title{Write a CAS Table to a CSV File}
\usage{
cas.write.csv2(
  CASTable,
  file = "",
  quote = TRUE,
  eol = "\\n",
  na = "NA",
  row.names = TRUE,
  fileEncoding = ""
)
}
\description{
This function is identical to \code{\link{cas.write.csv}}
except that it wraps the \R \code{write.csv2}  function. 
The \code{write.csv2} function uses a comma 
for the decimal separator and a semicolon for the field
delimiter.
}
% Copyright SAS Institute
