% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{casRetrieve}
\alias{casRetrieve}
\title{Utility function that is run from several exported functions.}
\usage{
casRetrieve(caz, ...)
}
\description{
The generated functions, such as cas.table.tableInfo and
cas.decisionTree.dtreeTrain, or runAction are better candidates
for use by programmers.
}
\keyword{internal}
% Copyright SAS Institute
