% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_tab.R
\docType{class}
\name{CASTable-class}
\alias{CASTable-class}
\alias{CASTable}
\title{CAS Table Object}
\arguments{
\item{conn}{A \code{\link{CAS}} object that represents the connection and
session on the server.}

\item{tname}{An optional \code{character} string for the table name.}

\item{caslib}{An optional \code{character} string that identifies the caslib for the
in-memory table. Specify this parameter to override the active caslib.}

\item{where}{An optional \code{character} string that specifies a filter for the
rows to process. The filter uses syntax that is specific to SAS.}

\item{orderby}{An optional \code{list} of column names. Rows are partitioned according
to the columns in the groupby parameter and then ordered according to
the values of the columns specified in this parameter.}

\item{groupby}{An optional \code{list} of column names. If you specify this parameter
when you load an in-memory table, then the table is partitioned by the
columns. If you specify this parameter when running an action, then
BY-groups are formed temporarily for the duration of the action.}

\item{gbmode}{An optional \code{character} string. Values are NOSORT (default) or
REDISTRIBUTE. See the CAS product documentation for more information.}

\item{computedOnDemand}{An optional \code{logical} flag that indicates whether to
the computed variables are created when the table is loaded (False) or to compute
them when an action begins (True).}

\item{computedVars}{An optional \code{character} string list that identifies the
name and optional information such as a format and label.}

\item{computedVarsProgram}{An optional \code{character} string list. Specify the
expression to use for computing each of the computed variables.}
}
\value{
\code{CASTable}
}
\description{
CASTable objects contain a reference to a data table loaded in a 
CAS server. This reference can also include filtering applied by
an expression, as well as additional computed columns. The filtering
and additional columns are only applied when using this object to
access data in the server-side table. In this way, CASTable objects
can be used like a client-side view of the table.
}
\section{Slots}{

\describe{
\item{\code{conn}}{A \code{\link{CAS}} object that represents the connection and
session on the server.}

\item{\code{tname}}{An optional \code{character} string for the table name.}

\item{\code{caslib}}{An optional \code{character} string that identifies the caslib for the
in-memory table. Specify this parameter to override the active caslib.}

\item{\code{where}}{An optional \code{character} string that specifies a filter for the
rows to process. The filter uses syntax that is specific to SAS.}

\item{\code{orderby}}{An optional \code{list} of column names. Rows are partitioned according
to the columns in the groupby parameter and then ordered according to
the values of the columns specified in this parameter.}

\item{\code{groupby}}{An optional \code{list} of column names. If you specify this parameter
when you load an in-memory table, then the table is partitioned by the
columns. If you specify this parameter when running an action, then
BY-groups are formed temporarily for the duration of the action.}

\item{\code{gbmode}}{An optional \code{character} string. Values are NOSORT (default) or
REDISTRIBUTE. See the CAS product documentation for more information.}

\item{\code{computedOnDemand}}{An optional \code{logical} flag that indicates whether to
the computed variables are created when the table is loaded (False) or to compute
them when an action begins (True).}

\item{\code{computedVars}}{An optional \code{character} string list that identifies the
name and optional information such as a format and label.}

\item{\code{computedVarsProgram}}{An optional \code{character} string list. Specify the
expression to use for computing each of the computed variables.}

\item{\code{names}}{An optional \code{list} of column names.}
}}

\seealso{
\code{as.CASTable}
}
