% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{cbind2.casTable}
\alias{cbind2.casTable}
\title{Combine CAS Tables by Columns}
\usage{
cbind2.casTable(x, y, ...)
}
\arguments{
\item{x}{\code{\link{CASTable}}.}

\item{y}{\code{\link{CASTable}}.}

\item{\dots}{Arguments that are passed to method arguments.}
}
\value{
\code{\link{CASTable}}
}
\description{
This is the implementation of \code{cbind} for in-memory tables.
}
\examples{
\dontrun{
cbind2(ct1, ct2)
cbind2(ct1[1:3], ct2$X1)
}

}
