% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{colMeans,CASTable-method}
\alias{colMeans,CASTable-method}
\title{Column Means}
\usage{
\S4method{colMeans}{CASTable}(x)
}
\arguments{
\item{x}{CASTable.}
}
\value{
vector
}
\description{
Column Means
}
\examples{
\dontrun{
colMeans(ct[1:4])
colMeans(ct$X1)
}
}
\seealso{
\code{cas.mean}
}
% Copyright SAS Institute
