% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\docType{methods}
\name{colSums,CASTable-method}
\alias{colSums,CASTable-method}
\title{Column Sums}
\usage{
\S4method{colSums}{CASTable}(x, na.rm = FALSE, dims = 1, ...)
}
\arguments{
\item{x}{CASTable.}
}
\value{
vector
}
\description{
Column Sums
}
\examples{
\dontrun{
colSums(ct[1:4])
colSums(ct$X1)
}
}
\seealso{
\code{cas.sum}
}
% Copyright SAS Institute
