% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{cov,CASTable-method}
\alias{cov,CASTable-method}
\title{Covariance}
\usage{
\S4method{cov}{CASTable}(x, y = NULL, use = "everything", method = c("pearson"))
}
\arguments{
\item{x}{CASTable.}
}
\value{
matrix
}
\description{
Unlike the \code{cov} function for data frames, 
this function does not support specifying the
method as "kendall" or "spearman." The results
for method "pearson" are returned.
}
\examples{
\dontrun{
cov(ct1)
cov(i2[1:4])
cov(i2[c('col1', 'col2', 'col3')])
cov(i2$col1, i2$col2)
}
}
% Copyright SAS Institute
