\name{CAS Actions}
\alias{cas.aStore.describe}
\alias{cas.aStore.download}
\alias{cas.aStore.score}
\alias{cas.aStore.upload}
\alias{cas.accessControl.assumeRole}
\alias{cas.accessControl.checkInAllObjects}
\alias{cas.accessControl.checkOutObject}
\alias{cas.accessControl.commitTransaction}
\alias{cas.accessControl.completeBackup}
\alias{cas.accessControl.createBackup}
\alias{cas.accessControl.deleteBWList}
\alias{cas.accessControl.dropRole}
\alias{cas.accessControl.isAuthorized}
\alias{cas.accessControl.isAuthorizedActions}
\alias{cas.accessControl.isAuthorizedColumns}
\alias{cas.accessControl.isAuthorizedTables}
\alias{cas.accessControl.isInRole}
\alias{cas.accessControl.listAcsActionSet}
\alias{cas.accessControl.listAcsData}
\alias{cas.accessControl.listAllPrincipals}
\alias{cas.accessControl.listMetadata}
\alias{cas.accessControl.operActionMd}
\alias{cas.accessControl.operActionSetMd}
\alias{cas.accessControl.operAdminMd}
\alias{cas.accessControl.operBWPaths}
\alias{cas.accessControl.operColumnMd}
\alias{cas.accessControl.operTableMd}
\alias{cas.accessControl.remAllAcsActionSet}
\alias{cas.accessControl.remAllAcsData}
\alias{cas.accessControl.repAllAcsAction}
\alias{cas.accessControl.repAllAcsActionSet}
\alias{cas.accessControl.repAllAcsCaslib}
\alias{cas.accessControl.repAllAcsColumn}
\alias{cas.accessControl.repAllAcsTable}
\alias{cas.accessControl.rollbackTransaction}
\alias{cas.accessControl.showRolesAllowed}
\alias{cas.accessControl.showRolesIn}
\alias{cas.accessControl.startTransaction}
\alias{cas.accessControl.statusTransaction}
\alias{cas.accessControl.updSomeAcsAction}
\alias{cas.accessControl.updSomeAcsActionSet}
\alias{cas.accessControl.updSomeAcsCaslib}
\alias{cas.accessControl.updSomeAcsColumn}
\alias{cas.accessControl.updSomeAcsTable}
\alias{cas.accessControl.whatCheckoutsExist}
\alias{cas.accessControl.whatIsEffective}
\alias{cas.aggregation.aggregate}
\alias{cas.bayesianNetClassifier.bnet}
\alias{cas.bglimmix.bglimmix}
\alias{cas.bioMedImage.buildSurface}
\alias{cas.boolRule.brScore}
\alias{cas.boolRule.brTrain}
\alias{cas.builtins.about}
\alias{cas.builtins.actionSetInfo}
\alias{cas.builtins.addNode}
\alias{cas.builtins.casCommon}
\alias{cas.builtins.echo}
\alias{cas.builtins.getLicenseInfo}
\alias{cas.builtins.getLicensedProductInfo}
\alias{cas.builtins.help}
\alias{cas.builtins.history}
\alias{cas.builtins.httpAddress}
\alias{cas.builtins.installActionSet}
\alias{cas.builtins.listNodes}
\alias{cas.builtins.loadActionSet}
\alias{cas.builtins.log}
\alias{cas.builtins.modifyQueue}
\alias{cas.builtins.ping}
\alias{cas.builtins.queryActionSet}
\alias{cas.builtins.queryName}
\alias{cas.builtins.reflect}
\alias{cas.builtins.refreshLicense}
\alias{cas.builtins.removeNode}
\alias{cas.builtins.serverStatus}
\alias{cas.builtins.shutdown}
\alias{cas.builtins.userInfo}
\alias{cas.cardinality.summarize}
\alias{cas.casclp.solveCsp}
\alias{cas.clustering.kClus}
\alias{cas.configuration.getServOpt}
\alias{cas.configuration.listServOpts}
\alias{cas.copula.copulaFit}
\alias{cas.copula.copulaSimulate}
\alias{cas.countreg.countregFitModel}
\alias{cas.dataDiscovery.profile}
\alias{cas.dataPreprocess.binning}
\alias{cas.dataPreprocess.catTrans}
\alias{cas.dataPreprocess.discretize}
\alias{cas.dataPreprocess.highCardinality}
\alias{cas.dataPreprocess.histogram}
\alias{cas.dataPreprocess.impute}
\alias{cas.dataPreprocess.kde}
\alias{cas.dataPreprocess.outlier}
\alias{cas.dataPreprocess.rustats}
\alias{cas.dataPreprocess.transform}
\alias{cas.dataStep.runCode}
\alias{cas.decisionTree.dtreeCode}
\alias{cas.decisionTree.dtreeMerge}
\alias{cas.decisionTree.dtreePrune}
\alias{cas.decisionTree.dtreeScore}
\alias{cas.decisionTree.dtreeSplit}
\alias{cas.decisionTree.dtreeTrain}
\alias{cas.decisionTree.forestCode}
\alias{cas.decisionTree.forestScore}
\alias{cas.decisionTree.forestTrain}
\alias{cas.decisionTree.gbtreeCode}
\alias{cas.decisionTree.gbtreeScore}
\alias{cas.decisionTree.gbtreeTrain}
\alias{cas.ds2.runDS2}
\alias{cas.elasticsearch.index}
\alias{cas.elasticsearch.sandIndex}
\alias{cas.espCluster.listservers}
\alias{cas.espCluster.startservers}
\alias{cas.factmac.factmac}
\alias{cas.fastknn.fastknn}
\alias{cas.fedSql.execDirect}
\alias{cas.freqTab.freqTab}
\alias{cas.gam.gampl}
\alias{cas.gam.gamplScore}
\alias{cas.glrm.hdpca}
\alias{cas.glrm.nnmf}
\alias{cas.gvarclus.gvarclus}
\alias{cas.hiddenMarkovModel.hmm}
\alias{cas.hyperGroup.hypergroup}
\alias{cas.image.augmentImages}
\alias{cas.image.compareImages}
\alias{cas.image.fetchImages}
\alias{cas.image.flattenImageTable}
\alias{cas.image.loadImages}
\alias{cas.image.matchImages}
\alias{cas.image.processImages}
\alias{cas.image.saveImages}
\alias{cas.image.summarizeImages}
\alias{cas.loadStreams.appendSnapshot}
\alias{cas.loadStreams.loadSnapshot}
\alias{cas.loadStreams.loadStream}
\alias{cas.loadStreams.mMetaData}
\alias{cas.localSearch.solveLso}
\alias{cas.mixed.blup}
\alias{cas.mixed.mixed}
\alias{cas.network.biconnectedComponents}
\alias{cas.network.centrality}
\alias{cas.network.clique}
\alias{cas.network.community}
\alias{cas.network.connectedComponents}
\alias{cas.network.core}
\alias{cas.network.cycle}
\alias{cas.network.path}
\alias{cas.network.reach}
\alias{cas.network.readGraph}
\alias{cas.network.shortestPath}
\alias{cas.network.summary}
\alias{cas.network.transitiveClosure}
\alias{cas.networkOptimization.LAP}
\alias{cas.networkOptimization.MCF}
\alias{cas.networkOptimization.MST}
\alias{cas.networkOptimization.linearAssignment}
\alias{cas.networkOptimization.minCostFlow}
\alias{cas.networkOptimization.minCut}
\alias{cas.networkOptimization.minSpanTree}
\alias{cas.networkOptimization.tsp}
\alias{cas.networkSocial.centrality}
\alias{cas.networkSocial.community}
\alias{cas.networkSocial.core}
\alias{cas.networkSocial.reach}
\alias{cas.neuralNet.annCode}
\alias{cas.neuralNet.annScore}
\alias{cas.neuralNet.annTrain}
\alias{cas.optML.basis}
\alias{cas.optML.lasso}
\alias{cas.optML.lsqr}
\alias{cas.optML.randmat}
\alias{cas.optML.svm}
\alias{cas.optMiner.tuneDecisionTree}
\alias{cas.optMiner.tuneFactMac}
\alias{cas.optMiner.tuneForest}
\alias{cas.optMiner.tuneGradientBoostTree}
\alias{cas.optMiner.tuneNeuralNet}
\alias{cas.optMiner.tuneSvm}
\alias{cas.optNetwork.LAP}
\alias{cas.optNetwork.MCF}
\alias{cas.optNetwork.MST}
\alias{cas.optNetwork.biconnectedComponents}
\alias{cas.optNetwork.clique}
\alias{cas.optNetwork.connectedComponents}
\alias{cas.optNetwork.cycle}
\alias{cas.optNetwork.linearAssignment}
\alias{cas.optNetwork.minCostFlow}
\alias{cas.optNetwork.minCut}
\alias{cas.optNetwork.minSpanTree}
\alias{cas.optNetwork.path}
\alias{cas.optNetwork.readGraph}
\alias{cas.optNetwork.shortestPath}
\alias{cas.optNetwork.summary}
\alias{cas.optNetwork.transitiveClosure}
\alias{cas.optNetwork.tsp}
\alias{cas.optimization.solveLp}
\alias{cas.optimization.solveMilp}
\alias{cas.optimization.solveQp}
\alias{cas.optimization.tuner}
\alias{cas.pca.eig}
\alias{cas.pca.itergs}
\alias{cas.pca.nipals}
\alias{cas.pca.randompca}
\alias{cas.percentile.assess}
\alias{cas.percentile.boxPlot}
\alias{cas.percentile.percentile}
\alias{cas.pls.pls}
\alias{cas.qlim.qlim}
\alias{cas.quantreg.quantreg}
\alias{cas.recommend.recomAls}
\alias{cas.recommend.recomAppend}
\alias{cas.recommend.recomCreate}
\alias{cas.recommend.recomDocDist}
\alias{cas.recommend.recomKnnScore}
\alias{cas.recommend.recomKnnTrain}
\alias{cas.recommend.recomMfScore}
\alias{cas.recommend.recomRateinfo}
\alias{cas.recommend.recomSample}
\alias{cas.recommend.recomSearchIndex}
\alias{cas.recommend.recomSearchQuery}
\alias{cas.recommend.recomSim}
\alias{cas.regression.genmod}
\alias{cas.regression.glm}
\alias{cas.regression.logistic}
\alias{cas.ruleMining.fpgrowth}
\alias{cas.ruleMining.taxonomy}
\alias{cas.sampling.oversample}
\alias{cas.sampling.srs}
\alias{cas.sampling.stratified}
\alias{cas.search.appendIndex}
\alias{cas.search.buildAutoComplete}
\alias{cas.search.buildIndex}
\alias{cas.search.deleteDocuments}
\alias{cas.search.getSchema}
\alias{cas.search.searchAggregate}
\alias{cas.search.searchAutocomplete}
\alias{cas.search.searchIndex}
\alias{cas.search.valueCount}
\alias{cas.sentimentAnalysis.applySent}
\alias{cas.sequence.pathing}
\alias{cas.session.addNodeStatus}
\alias{cas.session.batchresults}
\alias{cas.session.endSession}
\alias{cas.session.fetchresult}
\alias{cas.session.flushresult}
\alias{cas.session.listSessions}
\alias{cas.session.listresults}
\alias{cas.session.metrics}
\alias{cas.session.sessionId}
\alias{cas.session.sessionName}
\alias{cas.session.sessionStatus}
\alias{cas.session.setLocale}
\alias{cas.session.timeout}
\alias{cas.sessionProp.addFmtLib}
\alias{cas.sessionProp.addFormat}
\alias{cas.sessionProp.deleteFormat}
\alias{cas.sessionProp.dropFmtLib}
\alias{cas.sessionProp.getSessOpt}
\alias{cas.sessionProp.listFmtLibs}
\alias{cas.sessionProp.listFmtRanges}
\alias{cas.sessionProp.listFmtSearch}
\alias{cas.sessionProp.listFmtValues}
\alias{cas.sessionProp.listSessOpts}
\alias{cas.sessionProp.promoteFmtLib}
\alias{cas.sessionProp.saveFmtLib}
\alias{cas.sessionProp.setFmtSearch}
\alias{cas.sessionProp.setSessOpt}
\alias{cas.severity.severity}
\alias{cas.severity.severityValidate}
\alias{cas.simple.correlation}
\alias{cas.simple.crossTab}
\alias{cas.simple.distinct}
\alias{cas.simple.freq}
\alias{cas.simple.groupBy}
\alias{cas.simple.mdSummary}
\alias{cas.simple.numRows}
\alias{cas.simple.paraCoord}
\alias{cas.simple.regression}
\alias{cas.simple.summary}
\alias{cas.simple.topK}
\alias{cas.simpleForecast.forecast}
\alias{cas.svDataDescription.svddTrain}
\alias{cas.svm.svmTrain}
\alias{cas.table.addCaslib}
\alias{cas.table.addTable}
\alias{cas.table.attribute}
\alias{cas.table.caslibInfo}
\alias{cas.table.columnInfo}
\alias{cas.table.deleteSource}
\alias{cas.table.dropCaslib}
\alias{cas.table.dropTable}
\alias{cas.table.fetch}
\alias{cas.table.fileInfo}
\alias{cas.table.loadDataSource}
\alias{cas.table.loadTable}
\alias{cas.table.partition}
\alias{cas.table.promote}
\alias{cas.table.queryCaslib}
\alias{cas.table.recordCount}
\alias{cas.table.save}
\alias{cas.table.shuffle}
\alias{cas.table.tableDetails}
\alias{cas.table.tableExists}
\alias{cas.table.tableInfo}
\alias{cas.table.update}
\alias{cas.table.upload}
\alias{cas.table.view}
\alias{cas.textMining.createtopic}
\alias{cas.textMining.tmMine}
\alias{cas.textMining.tmScore}
\alias{cas.textMining.tmSvd}
\alias{cas.textParse.tpAccumulate}
\alias{cas.textParse.tpParse}
\alias{cas.textParse.validateCategory}
\alias{cas.textRuleDevelop.compileCategory}
\alias{cas.textRuleDevelop.compileConcept}
\alias{cas.textRuleDevelop.saveTableToDisk}
\alias{cas.textRuleDevelop.validateCategory}
\alias{cas.textRuleDevelop.validateConcept}
\alias{cas.textRuleDiscover.termMap}
\alias{cas.textRuleScore.applyCategory}
\alias{cas.textRuleScore.applyConcept}
\alias{cas.textRuleScore.loadTableFromDisk}
\alias{cas.textSummarization.textSummarize}
\alias{cas.textTopic.tmCreateTopic}
\alias{cas.textUtil.scoreAstore}
\alias{cas.textUtil.tmAstore}
\alias{cas.textUtil.tmCooccur}
\alias{cas.textUtil.tmfindSimilar}
\alias{cas.timeData.forecast}
\alias{cas.timeData.runTimeCode}
\alias{cas.timeData.timeSeries}
\alias{cas.timeFrequency.stft}
\alias{cas.timeFrequency.window}
\alias{cas.transpose.transpose}
\alias{cas.tsReconcile.reconcileTwoLevels}
\alias{cas.tsinfo.getinfo}
\alias{cas.varReduce.super}
\alias{cas.varReduce.unsuper}
\title{Common Page for CAS Actions}
\usage{
cas.\emph{actionSet}.\emph{action}(CASorCASTab, parameters...)
}
\arguments{
\item{CASorCASTab}{An instance of a \code{\link{CAS}} object that 
represents a connection and CAS session, or an instance of a 
\code{\link{CASTable}}.}
\item{parameters}{Actions accept a series of parameters in 
key=value pair format. The parmeters are action-specific. See
the product documentation.}
}
\description{
When you connect to SAS Cloud Analytic Services (CAS), the SWAT
package software generates an \R function for each CAS
action that is available on the server.
}
\section{Examples}{
The following two functions are generated and correspond to the
table.tableInfo and simple.summary actions:

\code{
cas.table.tableInfo(irisct)
}

\code{
cas.simple.summary(irisct)
}
}

\section{Product Documentation}{
For a list of all the CAS actions that are available with 
SAS Visual Analytics, SAS Visual Statistics, and SAS Visual
Data Mining and Machine Learning, see the following URL:

\href{http://documentation.sas.com/?cdcId=vdmmlcdc&cdcVersion=8.11
&docsetId=allprodsactions&docsetTarget=actionsByName.htm}{SAS Viya
3.2 Programming: Actions and Action Sets by Name and Product}

The preceding URL applies to the SAS Viya 3.2 release. For
the latest product documentation for SAS Viya, see 
\href{http://support.sas.com/documentation/onlinedoc/viya/index.html}{
Documentation for SAS Viya}.
}
% Copyright SAS Institute
