% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R
\name{is.castable}
\alias{is.castable}
\title{Test if an Object is a CAS Table}
\usage{
is.castable(x)
}
\arguments{
\item{x}{An \code{R} object.}
}
\value{
boolean
}
\description{
Test if an Object is a CAS Table
}
\examples{
\dontrun{
is.castable(ct1)     # TRUE
is.castable(iris)    # FALSE
}
}
% Copyright SAS Institute
