% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{listActionParms}
\alias{listActionParms}
\title{List CAS Action Parameters by Name}
\usage{
listActionParms(conn, actn)
}
\arguments{
\item{conn}{An instance of a CAS object that represents
a connection and CAS session.}

\item{actn}{A \code{string} value that specifies
the action name. You can specify the following forms:
\itemize{
  \item \emph{action-set.action-name}
  \item \emph{action-name}
  \item cas.\emph{action-set.action-name}
}
The third form matches the generated functions for
the CAS actions.}
}
\description{
This function displays the parameters for the specified
CAS action.
}
\examples{
\dontrun{
# specify the action set and action name
listActionParms(conn, actn="simple.summary")

# fetch is in the table action set
listActionParms(s, actn="fetch")

# specify the generated function name
listActionParms(s, cas.regression.logistic)
}
}
% Copyright SAS Institute
