% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\docType{methods}
\name{max,CASTable-method}
\alias{max,CASTable-method}
\title{Maximum Values}
\usage{
\S4method{max}{CASTable}(x)
}
\arguments{
\item{x}{CASTable.}
}
\value{
scalar
}
\description{
Maximum Values
}
\examples{
\dontrun{
max(ct1$X1)
max(ct1[1:2])
}
}
\seealso{
\code{cas.max}
}
% Copyright SAS Institute
