% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\docType{methods}
\name{mean,CASTable-method}
\alias{mean,CASTable-method}
\title{Mean Value for a Single Column}
\usage{
\S4method{mean}{CASTable}(x, ...)
}
\arguments{
\item{x}{CASTable.}
}
\value{
scalar
}
\description{
Returns the mean value for the specified column in 
the input table.
}
\examples{
\dontrun{
mean(ct1$x1)
mean(ct1[2])
}
}
\seealso{
\code{cas.mean}
}
% Copyright SAS Institute
