% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{median,CASTable-method}
\alias{median,CASTable-method}
\title{Median Values}
\usage{
\S4method{median}{CASTable}(x, na.rm = FALSE)
}
\arguments{
\item{x}{\code{\link{CASTable}} object.}

\item{na.rm}{Should missing values be omitted from the calculations?}
}
\value{
Named numeric vector
}
\description{
Median Values
}
\examples{
\dontrun{
median(ct1$x1)
median(ct1[1:4])
}
}
\seealso{
\code{cas.median}
}
