% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{min,CASTable-method}
\alias{min,CASTable-method}
\title{Minimum Value}
\usage{
\S4method{min}{CASTable}(x)
}
\arguments{
\item{x}{CASTable.}
}
\value{
scalar
}
\description{
Minimum Value
}
\examples{
\dontrun{
min(ct1$X1)
min(ct1[1:2])
}
}
\seealso{
\code{cas.min}
}
% Copyright SAS Institute
