% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAStab.R
\docType{methods}
\name{names,CASTable-method}
\alias{names,CASTable-method}
\title{Names of a CAS Table}
\usage{
\S4method{names}{CASTable}(x)
}
\arguments{
\item{x}{A CASTable object.}
}
\value{
vector
}
\description{
Returns the list of column names for the in-memory
table that is referenced by the \code{\link{CASTable}} object.
}
\examples{
\dontrun{
names(ct1)
}
}
% Copyright SAS Institute
