% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{rbind2.CASTable}
\alias{rbind2.CASTable}
\alias{rbind2,CASTable,ANY-method}
\title{Combine CAS Tables by Rows}
\usage{
rbind2.CASTable(x, y, ...)

\S4method{rbind2}{CASTable,ANY}(x, y, ...)
}
\arguments{
\item{x}{\code{\link{CASTable}}.}

\item{y}{\code{\link{CASTable}}.}

\item{\dots}{Arguments that are passed to method arguments.}
}
\value{
\code{\link{CASTable}}
}
\description{
This is the implementation of \code{rbind} for in-memory tables.
}
\examples{
\dontrun{
rbind(ct1, ct2)
rbind(ct1, ct2, ct3)
}

}
