% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\docType{methods}
\name{summary,CASTable-method}
\alias{summary,CASTable-method}
\title{Summary Statistics}
\usage{
\S4method{summary}{CASTable}(object, maxsum = 7, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{CASTable.}
}
\value{
table
}
\description{
Returns simple descriptive statistics.
}
\examples{
\dontrun{
summary(ct1)
summary(ct1[1:4])
}
}
% Copyright SAS Institute
