% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_df.R
\name{to.casDataFrame}
\alias{to.casDataFrame}
\title{Convert a CAS Table to a CAS Data Frame (Download)}
\usage{
to.casDataFrame(x, obs = 32768)
}
\arguments{
\item{x}{The CAS table data to download.}

\item{obs}{Optional integer indicating the maximum number of rows
of data to download.}
}
\value{
Returns a casDataFrame object that contains
        a copy of the in-memory data.
}
\description{
Downloads the in-memory table that is referenced by
the CASTable object and stores it as a casDataFrame
in R. This function is used primarily by the package
to store the results of a CAS action.
}
\examples{
\dontrun{
cdf <- to.CASDataFrame(CASTable)
}

}
