% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASdf.R
\name{to.data.frame}
\alias{to.data.frame}
\title{Convert a CAS data frame to an \R Data Frame}
\usage{
to.data.frame(cdf)
}
\arguments{
\item{cdf}{The casDataFrame to convert}
}
\value{
Data Frame
}
\description{
This function returns the \R data.frame object that is encapsulated
in a casDataFrame. The CAS metadata is not available in the returned
data.frame.
}
\examples{
\dontrun{
df3 = to.data.frame(cdf)
}
}
% Copyright SAS Institute
