% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userReadSwatOutput.R
\name{userReadSwatOutput}
\alias{userReadSwatOutput}
\title{User-defined function for reading SWAT/SWAT+ output - DO NOT USE AS IT IS NOW}
\usage{
userReadSwatOutput()
}
\value{
a list of simulated data
}
\description{
This is a user-defined function for reading SWAT/SWAt+ outputs, do not use
as it is now, please modify the source code of this function if you use
according to your need and then install RSWAT app again from your local
folder to update this function.DO NOT CHANGE the name of this function
(mus be userObjFunction). ONLY CHANGE the code at the place where indicates
"START TO CHANGE FROM HERE" and STOP at "END OF CHANGE". This functions
must NOT receive any input. By default, you are NOW INSIDE the TxtInOut folder
and you can read any output files there by just using the file name.
RSWAT will use this function to get SWAT outputs (from the respective core)
after each model run and save it in combination with other outputs.

Please TEST your function before running R-SWAT
}
\examples{

\donttest{
userReadSwatOutput()
}

}
